Crate rustdoc

Unstable (rustdoc)

Modules

clean 	 [Unstable]
This module contains the "cleaned" pieces of the AST, and the functions that clean them.

core 	 [Unstable]

doctree 	 [Unstable]
This module is used to store stuff from Rust's AST in a more convenient manner (and with prettier names) before cleaning.

externalfiles 	 [Unstable]

fold 	 [Unstable]

html 	 [Unstable]

markdown 	 [Unstable]

passes 	 [Unstable]

plugins 	 [Unstable]

test 	 [Unstable]

visit_ast 	 [Unstable]
Rust AST Visitor. Extracts useful information and massages it into a form usable for clean

visit_lib 	 [Unstable]

Functions

main 	 [Unstable]

main_args 	 [Unstable]

opts 	 [Unstable]

usage 	 [Unstable]