Module rustdoc::clean

Unstable (rustdoc)

This module contains the "cleaned" pieces of the AST, and the functions that clean them.

Reexports

pub use self::Type::*;

pub use self::PrimitiveType::*;

pub use self::TypeKind::*;

pub use self::VariantKind::*;

pub use self::Mutability::*;

pub use self::Import::*;

pub use self::ItemEnum::*;

pub use self::Attribute::*;

pub use self::TyParamBound::*;

pub use self::SelfTy::*;

pub use self::FunctionRetTy::*;

pub use self::Visibility::*;

Modules

inline 	 [Unstable]
Support for inlining external documentation into the current AST.

Structs

Argument 	 [Unstable]

Arguments 	 [Unstable]

BareFunctionDecl 	 [Unstable]

Constant 	 [Unstable]

Crate 	 [Unstable]

DefaultImpl 	 [Unstable]

Deprecation 	 [Unstable]

Enum 	 [Unstable]

ExternalCrate 	 [Unstable]

FnDecl 	 [Unstable]

Function 	 [Unstable]

Generics 	 [Unstable]

Impl 	 [Unstable]

ImportSource 	 [Unstable]

Item 	 [Unstable]
Anything with a source location and set of attributes and, optionally, a name. That is, anything that can be documented. This doesn't correspond directly to the AST's concept of an item; it's a strict superset.

Lifetime 	 [Unstable]

Macro 	 [Unstable]

Method 	 [Unstable]

Module 	 [Unstable]

Path 	 [Unstable]

PathSegment 	 [Unstable]

PolyTrait 	 [Unstable]
A trait reference, which may have higher ranked lifetimes.

Span 	 [Unstable]

Stability 	 [Unstable]

Static 	 [Unstable]

Struct 	 [Unstable]

Trait 	 [Unstable]

TyMethod 	 [Unstable]

TyParam 	 [Unstable]

TypeBinding 	 [Unstable]
An equality constraint on an associated type, e.g. A=Bar in Foo<A=Bar>

Typedef 	 [Unstable]

Variant 	 [Unstable]

VariantStruct 	 [Unstable]
This is a more limited form of the standard Struct, different in that it lacks the things most items have (name, id, parameterization). Found only as a variant in an enum.

ViewListIdent 	 [Unstable]

Enums

Attribute 	 [Unstable]

FunctionRetTy 	 [Unstable]

ImplPolarity 	 [Unstable]

Import 	 [Unstable]

ItemEnum 	 [Unstable]

Mutability 	 [Unstable]

PathParameters 	 [Unstable]

PrimitiveType 	 [Unstable]

SelfTy 	 [Unstable]

TyParamBound 	 [Unstable]

Type 	 [Unstable]
A representation of a Type suitable for hyperlinking purposes. Ideally one can get the original type out of the AST/TyCtxt given one of these, if more information is needed. Most importantly it does not preserve mutability or boxes.

TypeKind 	 [Unstable]

VariantKind 	 [Unstable]

Visibility 	 [Unstable]

WherePredicate 	 [Unstable]

Traits

Attributes 	 [Unstable]

Clean 	 [Unstable]

GetDefId 	 [Unstable]