Enum rustdoc::clean::cfg::Cfg

pub enum Cfg {
    True,
    False,
    Cfg(Symbol, Option<Symbol>),
    Not(Box<Cfg>),
    Any(Vec<Cfg>),
    All(Vec<Cfg>),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Variants
True

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Accepts all configurations.

False

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Denies all configurations.

Cfg(Symbol, Option<Symbol>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A generic configration option, e.g. test or target_os = "linux".

Not(Box<Cfg>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Negate a configuration requirement, i.e. not(x).

Any(Vec<Cfg>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Union of a list of configuration requirements, i.e. any(...).

All(Vec<Cfg>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Intersection of a list of configuration requirements, i.e. all(...).

Methods

impl Cfg

fn parse(cfg: &MetaItem) -> Result<Cfg, InvalidCfgError>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Parses a MetaItem into a Cfg.

The MetaItem should be the content of the #[cfg(...)], e.g. unix or target_os = "redox".

If the content is not properly formatted, it will return an error indicating what and where the error is.

fn matches(&self, parse_sess: &ParseSess, features: Option<&Features>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Checks whether the given configuration can be matched in the current session.

Equivalent to attr::cfg_matches.