Module rustdoc::clean::inline

Unstable (rustdoc)

Support for inlining external documentation into the current AST.

Functions

build_external_trait 	 [Unstable]

build_impl 	 [Unstable]

build_impls 	 [Unstable]

load_attrs 	 [Unstable]

record_extern_fqn 	 [Unstable]
Record an external fully qualified name in the external_paths cache.

try_inline 	 [Unstable]
Attempt to inline the definition of a local node id into this AST.