Module rustdoc::clean::render_macro_matchers

Functions

print_tt 🔒

print_tts 🔒

render_macro_matcher 🔒

Render a macro matcher in a format suitable for displaying to the user as part of an item declaration.

snippet_equal_to_token 🔒

Find the source snippet for this token’s Span, reparse it, and return the snippet if the reparsed TokenTree matches the argument TokenTree.

usually_needs_space_between_keyword_and_open_delim 🔒