Struct rustdoc::clean::Static

pub struct Static {
    pub type_: Type,
    pub mutability: Mutability,
    pub expr: String,
}

Unstable (rustdoc)

Fields

type_
Unstable (rustdoc)

mutability
Unstable (rustdoc)

expr
Unstable (rustdoc)

It's useful to have the value of a static documented, but I have no desire to represent expressions (that'd basically be all of the AST, which is huge!). So, have a string.