Struct rustdoc::clean::VariantStruct

pub struct VariantStruct {
    pub struct_type: StructType,
    pub fields: Vec<Item>,
    pub fields_stripped: bool,
}

Unstable (rustdoc)

This is a more limited form of the standard Struct, different in that it lacks the things most items have (name, id, parameterization). Found only as a variant in an enum.

Fields
struct_type: StructType
Unstable (rustdoc)
fields: Vec<Item>
Unstable (rustdoc)
fields_stripped: bool
Unstable (rustdoc)