Trait rustdoc::clean::Clean

pub trait Clean<T> {
    fn clean(&self, cx: &DocContext) -> T;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn clean(&self, cx: &DocContext) -> T

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?