Enum rustdoc::clean::types::Type

pub(crate) enum Type {

    Path {
        path: Path,
    },
    DynTrait(Vec<PolyTrait>, Option<Lifetime>),
    Generic(Symbol),
    Primitive(PrimitiveType),
    BareFunction(Box<BareFunctionDecl>),
    Tuple(Vec<Type>),
    Slice(Box<Type>),
    Array(Box<Type>, String),
    RawPointer(Mutability, Box<Type>),
    BorrowedRef {
        lifetime: Option<Lifetime>,
        mutability: Mutability,
        type_: Box<Type>,
    },
    QPath(Box<QPathData>),
    Infer,
    ImplTrait(Vec<GenericBound>),

}

Rustdoc’s representation of types, mostly based on the hir::Ty.

Variants

Path

Fields

path: Path

A named type, which could be a trait.

This is mostly Rustdoc’s version of hir::Path. It has to be different because Rustdoc’s PathSegment can contain cleaned generics.

DynTrait(Vec<PolyTrait>, Option<Lifetime>)

A dyn Trait object: dyn for<'a> Trait<'a> + Send + 'static

Generic(Symbol)

A type parameter.

Primitive(PrimitiveType)

A primitive (aka, builtin) type.

BareFunction(Box<BareFunctionDecl>)

A function pointer: extern "ABI" fn(...) -> ...

Tuple(Vec<Type>)

A tuple type: (i32, &str).

Slice(Box<Type>)

A slice type (does not include the &): [i32]

Array(Box<Type>, String)

An array type.

The String field is a stringified version of the array’s length parameter.

RawPointer(Mutability, Box<Type>)

A raw pointer type: *const i32, *mut i32

BorrowedRef

Fields

lifetime: Option<Lifetime>

mutability: Mutability

type_: Box<Type>

A reference type: &i32, &'a mut Foo

QPath(Box<QPathData>)

A qualified path to an associated item: <Type as Trait>::Name

Infer

A type that is inferred: _

ImplTrait(Vec<GenericBound>)

An impl Trait: impl TraitA + TraitB + ...

Implementations

impl Type

pub(crate) fn without_borrowed_ref(&self) -> &Type

When comparing types for equality, it can help to ignore & wrapping.

pub(crate) fn is_same(&self, other: &Self, cache: &Cache) -> bool

Check if two types are “potentially the same”. This is different from Eq, because it knows that things like Placeholder are possible matches for everything.

pub(crate) fn primitive_type(&self) -> Option<PrimitiveType>

pub(crate) fn is_assoc_ty(&self) -> bool

Checks if this is a T::Name path for an associated type.

pub(crate) fn is_self_type(&self) -> bool

pub(crate) fn generics(&self) -> Option<Vec<&Type>>

pub(crate) fn is_full_generic(&self) -> bool

pub(crate) fn is_impl_trait(&self) -> bool

pub(crate) fn projection(&self) -> Option<(&Type, DefId, PathSegment)>

fn inner_def_id(&self, cache: Option<&Cache>) -> Option<DefId>

pub(crate) fn def_id(&self, cache: &Cache) -> Option<DefId>

Use this method to get the DefId of a clean AST node, including PrimitiveTypes.

impl Type

pub(crate) fn print<'b, 'a: 'b, 'tcx: 'a>(
&'a self,
cx: &'a Context<'tcx>
) -> impl Display + 'b + Captures<'tcx>