rustdoc

Module config

Structs

Options 🔒

Configuration options for rustdoc.

PathToParts 🔒

Path directly to crate-info file.

RenderOptions 🔒

Configuration options for the HTML page-creation process.

ShouldMerge 🔒

Controls merging of cross-crate information

Enums

EmitType 🔒

InputMode 🔒

Either an input crate, markdown file, or nothing (–merge=finalize).

ModuleSorting 🔒

OutputFormat 🔒

Functions

check_deprecated_options 🔒

Prints deprecation warnings for deprecated options

make_input 🔒

Create the input (string or file path)

markdown_input 🔒

Returns true if the file given as self.input is a Markdown file.

parse_extern_html_roots 🔒

Extracts --extern-html-root-url arguments from matches and returns a map of crate names to the given URLs. If an --extern-html-root-url argument was ill-formed, returns an error describing the issue.

parse_include_parts_dir 🔒

Reports error if –include-parts-dir / crate-info is not a file

parse_merge 🔒

Extracts read_rendered_cci and write_rendered_cci from command line arguments, or reports an error if an invalid option was provided

parse_remap_path_prefix 🔒