rustdoc::core

Function new_dcx

pub(crate) fn new_dcx(
    error_format: ErrorOutputType,
    source_map: Option<Arc<SourceMap>>,
    diagnostic_width: Option<usize>,
    unstable_opts: &UnstableOptions,
) -> DiagCtxt

Creates a new DiagCtxt that can be used to emit warnings and errors.

If the given error_format is ErrorOutputType::Json and no SourceMap is given, a new one will be created for the DiagCtxt.