Struct rustdoc::core::DocContext

pub struct DocContext<'a, 'tcx: 'a> {
    pub tcx: TyCtxt<'a, 'tcx, 'tcx>,
    pub populated_all_crate_impls: Cell<bool>,
    pub access_levels: RefCell<AccessLevels<DefId>>,
    pub renderinfo: RefCell<RenderInfo>,
    pub external_traits: RefCell<FxHashMap<DefId, Trait>>,
    pub ty_substs: RefCell<FxHashMap<Def, Type>>,
    pub lt_substs: RefCell<FxHashMap<NodeId, Lifetime>>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
tcx: TyCtxt<'a, 'tcx, 'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

populated_all_crate_impls: Cell<bool>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

access_levels: RefCell<AccessLevels<DefId>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Later on moved into clean::Crate

renderinfo: RefCell<RenderInfo>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Later on moved into html::render::CACHE_KEY

external_traits: RefCell<FxHashMap<DefId, Trait>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Later on moved through clean::Crate into html::render::CACHE_KEY

ty_substs: RefCell<FxHashMap<Def, Type>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Table type parameter definition -> substituted type

lt_substs: RefCell<FxHashMap<NodeId, Lifetime>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Table node id of lifetime parameter definition -> substituted lifetime

Methods

impl<'a, 'tcx> DocContext<'a, 'tcx>

fn sess(&self) -> &Session

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn enter_alias<F, R>(
&self,
ty_substs: FxHashMap<Def, Type>,
lt_substs: FxHashMap<NodeId, Lifetime>,
f: F
) -> R where
F: FnOnce() -> R,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Call the closure with the given parameters set as the substitutions for a type alias' RHS.