Struct rustdoc::core::UnstableOptions

pub(crate) struct UnstableOptions {

    pub allow_features: Option<Vec<String, Global>>,
    pub always_encode_mir: bool,
    pub asm_comments: bool,
    pub assert_incr_state: Option<String>,
    pub assume_incomplete_release: bool,
    pub binary_dep_depinfo: bool,
    pub box_noalias: bool,
    pub branch_protection: Option<BranchProtection>,
    pub cf_protection: CFProtection,
    pub codegen_backend: Option<String>,
    pub combine_cgu: bool,
    pub crate_attr: Vec<String, Global>,
    pub debug_info_for_profiling: bool,
    pub debug_macros: bool,
    pub deduplicate_diagnostics: bool,
    pub dep_info_omit_d_target: bool,
    pub dep_tasks: bool,
    pub dont_buffer_diagnostics: bool,
    pub drop_tracking: bool,
    pub drop_tracking_mir: bool,
    pub dual_proc_macros: bool,
    pub dump_dep_graph: bool,
    pub dump_drop_tracking_cfg: Option<String>,
    pub dump_mir: Option<String>,
    pub dump_mir_dataflow: bool,
    pub dump_mir_dir: String,
    pub dump_mir_exclude_pass_number: bool,
    pub dump_mir_graphviz: bool,
    pub dump_mir_spanview: Option<MirSpanview>,
    pub dump_mono_stats: SwitchWithOptPath,
    pub dump_mono_stats_format: DumpMonoStatsFormat,
    pub dump_solver_proof_tree: DumpSolverProofTree,
    pub dump_solver_proof_tree_use_cache: Option<bool>,
    pub dwarf_version: Option<u32>,
    pub dylib_lto: bool,
    pub emit_stack_sizes: bool,
    pub emit_thin_lto: bool,
    pub export_executable_symbols: bool,
    pub extra_const_ub_checks: bool,
    pub fewer_names: Option<bool>,
    pub flatten_format_args: bool,
    pub force_unstable_if_unmarked: bool,
    pub fuel: Option<(String, u64)>,
    pub function_sections: Option<bool>,
    pub future_incompat_test: bool,
    pub gcc_ld: Option<LdImpl>,
    pub graphviz_dark_mode: bool,
    pub graphviz_font: String,
    pub hir_stats: bool,
    pub human_readable_cgu_names: bool,
    pub identify_regions: bool,
    pub incremental_ignore_spans: bool,
    pub incremental_info: bool,
    pub incremental_relative_spans: bool,
    pub incremental_verify_ich: bool,
    pub inline_in_all_cgus: Option<bool>,
    pub inline_llvm: bool,
    pub inline_mir: Option<bool>,
    pub inline_mir_hint_threshold: Option<usize>,
    pub inline_mir_threshold: Option<usize>,
    pub input_stats: bool,
    pub instrument_coverage: Option<InstrumentCoverage>,
    pub instrument_mcount: bool,
    pub instrument_xray: Option<InstrumentXRay>,
    pub keep_hygiene_data: bool,
    pub layout_seed: Option<u64>,
    pub link_directives: bool,
    pub link_native_libraries: bool,
    pub link_only: bool,
    pub llvm_plugins: Vec<String, Global>,
    pub llvm_time_trace: bool,
    pub location_detail: LocationDetail,
    pub ls: bool,
    pub macro_backtrace: bool,
    pub maximal_hir_to_mir_coverage: bool,
    pub merge_functions: Option<MergeFunctions>,
    pub meta_stats: bool,
    pub mir_emit_retag: bool,
    pub mir_enable_passes: Vec<(String, bool), Global>,
    pub mir_include_spans: bool,
    pub mir_keep_place_mention: bool,
    pub mir_opt_level: Option<usize>,
    pub move_size_limit: Option<usize>,
    pub mutable_noalias: bool,
    pub nll_facts: bool,
    pub nll_facts_dir: String,
    pub no_analysis: bool,
    pub no_codegen: bool,
    pub no_generate_arange_section: bool,
    pub no_jump_tables: bool,
    pub no_leak_check: bool,
    pub no_link: bool,
    pub no_parallel_llvm: bool,
    pub no_profiler_runtime: bool,
    pub no_unique_section_names: bool,
    pub normalize_docs: bool,
    pub oom: OomStrategy,
    pub osx_rpath_install_name: bool,
    pub packed_bundled_libs: bool,
    pub panic_abort_tests: bool,
    pub panic_in_drop: PanicStrategy,
    pub parse_only: bool,
    pub perf_stats: bool,
    pub plt: Option<bool>,
    pub polonius: bool,
    pub polymorphize: bool,
    pub pre_link_arg: (),
    pub pre_link_args: Vec<String, Global>,
    pub precise_enum_drop_elaboration: bool,
    pub print_codegen_stats: bool,
    pub print_fuel: Option<String>,
    pub print_llvm_passes: bool,
    pub print_mono_items: Option<String>,
    pub print_type_sizes: bool,
    pub print_vtable_sizes: bool,
    pub proc_macro_backtrace: bool,
    pub proc_macro_execution_strategy: ProcMacroExecutionStrategy,
    pub profile: bool,
    pub profile_closures: bool,
    pub profile_emit: Option<PathBuf>,
    pub profile_sample_use: Option<PathBuf>,
    pub profiler_runtime: String,
    pub query_dep_graph: bool,
    pub randomize_layout: bool,
    pub relax_elf_relocations: Option<bool>,
    pub relro_level: Option<RelroLevel>,
    pub remap_cwd_prefix: Option<PathBuf>,
    pub remark_dir: Option<PathBuf>,
    pub report_delayed_bugs: bool,
    pub sanitizer: SanitizerSet,
    pub sanitizer_cfi_canonical_jump_tables: Option<bool>,
    pub sanitizer_cfi_generalize_pointers: Option<bool>,
    pub sanitizer_cfi_normalize_integers: Option<bool>,
    pub sanitizer_memory_track_origins: usize,
    pub sanitizer_recover: SanitizerSet,
    pub saturating_float_casts: Option<bool>,
    pub self_profile: SwitchWithOptPath,
    pub self_profile_counter: String,
    pub self_profile_events: Option<Vec<String, Global>>,
    pub share_generics: Option<bool>,
    pub show_span: Option<String>,
    pub simulate_remapped_rust_src_base: Option<PathBuf>,
    pub span_debug: bool,
    pub span_free_formats: bool,
    pub split_dwarf_inlining: bool,
    pub split_dwarf_kind: SplitDwarfKind,
    pub split_lto_unit: Option<bool>,
    pub src_hash_algorithm: Option<SourceFileHashAlgorithm>,
    pub stack_protector: StackProtector,
    pub staticlib_allow_rdylib_deps: bool,
    pub staticlib_prefer_dynamic: bool,
    pub strict_init_checks: bool,
    pub strip: Strip,
    pub symbol_mangling_version: Option<SymbolManglingVersion>,
    pub teach: bool,
    pub temps_dir: Option<String>,
    pub terminal_urls: TerminalUrl,
    pub thinlto: Option<bool>,
    pub thir_unsafeck: bool,
    pub threads: usize,
    pub time_llvm_passes: bool,
    pub time_passes: bool,
    pub time_passes_format: TimePassesFormat,
    pub tiny_const_eval_limit: bool,
    pub tls_model: Option<TlsModel>,
    pub trace_macros: bool,
    pub track_diagnostics: bool,
    pub trait_solver: TraitSolver,
    pub translate_additional_ftl: Option<PathBuf>,
    pub translate_directionality_markers: bool,
    pub translate_lang: Option<LanguageIdentifier>,
    pub translate_remapped_path_to_local_path: bool,
    pub trap_unreachable: Option<bool>,
    pub treat_err_as_bug: Option<NonZeroUsize>,
    pub trim_diagnostic_paths: bool,
    pub tune_cpu: Option<String>,
    pub ui_testing: bool,
    pub uninit_const_chunk_threshold: usize,
    pub unleash_the_miri_inside_of_you: bool,
    pub unpretty: Option<String>,
    pub unsound_mir_opts: bool,
    pub unstable_options: bool,
    pub use_ctors_section: Option<bool>,
    pub validate_mir: bool,
    pub verbose: bool,
    pub verify_llvm_ir: bool,
    pub virtual_function_elimination: bool,
    pub wasi_exec_model: Option<WasiExecModel>,
    pub write_long_types_to_disk: bool,

}

Fields
allow_features: Option<Vec<String, Global>>always_encode_mir: boolasm_comments: boolassert_incr_state: Option<String>assume_incomplete_release: boolbinary_dep_depinfo: boolbox_noalias: boolbranch_protection: Option<BranchProtection>cf_protection: CFProtectioncodegen_backend: Option<String>combine_cgu: boolcrate_attr: Vec<String, Global>debug_info_for_profiling: booldebug_macros: booldeduplicate_diagnostics: booldep_info_omit_d_target: booldep_tasks: booldont_buffer_diagnostics: booldrop_tracking: booldrop_tracking_mir: booldual_proc_macros: booldump_dep_graph: booldump_drop_tracking_cfg: Option<String>dump_mir: Option<String>dump_mir_dataflow: booldump_mir_dir: Stringdump_mir_exclude_pass_number: booldump_mir_graphviz: booldump_mir_spanview: Option<MirSpanview>dump_mono_stats: SwitchWithOptPathdump_mono_stats_format: DumpMonoStatsFormatdump_solver_proof_tree: DumpSolverProofTreedump_solver_proof_tree_use_cache: Option<bool>dwarf_version: Option<u32>dylib_lto: boolemit_stack_sizes: boolemit_thin_lto: boolexport_executable_symbols: boolextra_const_ub_checks: boolfewer_names: Option<bool>flatten_format_args: boolforce_unstable_if_unmarked: boolfuel: Option<(String, u64)>function_sections: Option<bool>future_incompat_test: boolgcc_ld: Option<LdImpl>graphviz_dark_mode: boolgraphviz_font: Stringhir_stats: boolhuman_readable_cgu_names: boolidentify_regions: boolincremental_ignore_spans: boolincremental_info: boolincremental_relative_spans: boolincremental_verify_ich: boolinline_in_all_cgus: Option<bool>inline_llvm: boolinline_mir: Option<bool>inline_mir_hint_threshold: Option<usize>inline_mir_threshold: Option<usize>input_stats: boolinstrument_coverage: Option<InstrumentCoverage>instrument_mcount: boolinstrument_xray: Option<InstrumentXRay>keep_hygiene_data: boollayout_seed: Option<u64>link_directives: boollink_native_libraries: boollink_only: boolllvm_plugins: Vec<String, Global>llvm_time_trace: boollocation_detail: LocationDetaills: boolmacro_backtrace: boolmaximal_hir_to_mir_coverage: boolmerge_functions: Option<MergeFunctions>meta_stats: boolmir_emit_retag: boolmir_enable_passes: Vec<(String, bool), Global>mir_include_spans: boolmir_keep_place_mention: boolmir_opt_level: Option<usize>move_size_limit: Option<usize>mutable_noalias: boolnll_facts: boolnll_facts_dir: Stringno_analysis: boolno_codegen: boolno_generate_arange_section: boolno_jump_tables: boolno_leak_check: boolno_link: boolno_parallel_llvm: boolno_profiler_runtime: boolno_unique_section_names: boolnormalize_docs: booloom: OomStrategyosx_rpath_install_name: boolpacked_bundled_libs: boolpanic_abort_tests: boolpanic_in_drop: PanicStrategyparse_only: boolperf_stats: boolplt: Option<bool>polonius: boolpolymorphize: boolpre_link_arg: ()pre_link_args: Vec<String, Global>precise_enum_drop_elaboration: boolprint_codegen_stats: boolprint_fuel: Option<String>print_llvm_passes: boolprint_mono_items: Option<String>print_type_sizes: boolprint_vtable_sizes: boolproc_macro_backtrace: boolproc_macro_execution_strategy: ProcMacroExecutionStrategyprofile: boolprofile_closures: boolprofile_emit: Option<PathBuf>profile_sample_use: Option<PathBuf>profiler_runtime: Stringquery_dep_graph: boolrandomize_layout: boolrelax_elf_relocations: Option<bool>relro_level: Option<RelroLevel>remap_cwd_prefix: Option<PathBuf>remark_dir: Option<PathBuf>report_delayed_bugs: boolsanitizer: SanitizerSetsanitizer_cfi_canonical_jump_tables: Option<bool>sanitizer_cfi_generalize_pointers: Option<bool>sanitizer_cfi_normalize_integers: Option<bool>sanitizer_memory_track_origins: usizesanitizer_recover: SanitizerSetsaturating_float_casts: Option<bool>self_profile: SwitchWithOptPathself_profile_counter: Stringself_profile_events: Option<Vec<String, Global>>

keep this in sync with the event filter names in librustc_data_structures/profiling.rs

share_generics: Option<bool>show_span: Option<String>simulate_remapped_rust_src_base: Option<PathBuf>span_debug: boolspan_free_formats: bool

o/w tests have closure@path

split_dwarf_inlining: boolsplit_dwarf_kind: SplitDwarfKindsplit_lto_unit: Option<bool>src_hash_algorithm: Option<SourceFileHashAlgorithm>stack_protector: StackProtectorstaticlib_allow_rdylib_deps: boolstaticlib_prefer_dynamic: boolstrict_init_checks: boolstrip: Stripsymbol_mangling_version: Option<SymbolManglingVersion>teach: booltemps_dir: Option<String>terminal_urls: TerminalUrlthinlto: Option<bool>thir_unsafeck: boolthreads: usize

We default to 1 here since we want to behave like a sequential compiler for now. This’ll likely be adjusted in the future. Note that -Zthreads=0 is the way to get the num_cpus behavior.

time_llvm_passes: booltime_passes: booltime_passes_format: TimePassesFormattiny_const_eval_limit: booltls_model: Option<TlsModel>trace_macros: booltrack_diagnostics: booltrait_solver: TraitSolvertranslate_additional_ftl: Option<PathBuf>translate_directionality_markers: booltranslate_lang: Option<LanguageIdentifier>translate_remapped_path_to_local_path: booltrap_unreachable: Option<bool>treat_err_as_bug: Option<NonZeroUsize>trim_diagnostic_paths: booltune_cpu: Option<String>ui_testing: booluninit_const_chunk_threshold: usizeunleash_the_miri_inside_of_you: boolunpretty: Option<String>unsound_mir_opts: boolunstable_options: bool

This name is kind of confusing: Most unstable options enable something themselves, while this just allows “normal” options to be feature-gated.

use_ctors_section: Option<bool>validate_mir: boolverbose: boolverify_llvm_ir: boolvirtual_function_elimination: boolwasi_exec_model: Option<WasiExecModel>write_long_types_to_disk: bool