Module rustdoc::doctest

Structs

Collector ğŸ”’

GlobalTestOptions ğŸ”’

Options that apply to all doctests in a crate or Markdown file (for rustdoc foo.md).

HirCollector ğŸ”’

UnusedExterns ğŸ”’

Enums

DirState ğŸ”’

TestFailure ğŸ”’

Documentation test failure modes.

Traits

Tester ğŸ”’

Functions

check_if_attr_is_complete ğŸ”’

make_test ğŸ”’

Transforms a test into code that can be compiled into a Rust binary, and returns the number of lines before the test code begins as well as if the output stream supports colors or not.

partition_source ğŸ”’

run ğŸ”’

run_test ğŸ”’

run_tests ğŸ”’

scrape_test_config ğŸ”’