rustdoc::doctest

Function run_test

fn run_test(
    doctest: RunnableDocTest,
    rustdoc_options: &Options,
    supports_color: bool,
    report_unused_externs: impl Fn(UnusedExterns),
) -> (Duration, Result<(), TestFailure>)

Execute a RunnableDoctest.

This is the function that calculates the compiler command line, invokes the compiler, then invokes the test or tests in a separate executable (if applicable).

Returns a tuple containing the Duration of the compilation and the Result of the test.