Module rustdoc::doctest::make

Logic for transforming the raw code given by the user into something actually runnable, e.g. by adding a main function if it doesn’t already exist.

Functions

check_for_main_and_extern_crate 🔒

check_if_attr_is_complete 🔒

make_test 🔒

Transforms a test into code that can be compiled into a Rust binary, and returns the number of lines before the test code begins as well as if the output stream supports colors or not.

partition_source 🔒