Trait rustdoc::fold::DocFolder

pub trait DocFolder: Sized {
    fn fold_item(&mut self, item: Item) -> Option<Item> { ... }

    fn fold_inner_recur(&mut self, inner: ItemEnum) -> ItemEnum { ... }

    fn fold_item_recur(&mut self, item: Item) -> Option<Item> { ... }

    fn fold_mod(&mut self, m: Module) -> Module { ... }

    fn fold_crate(&mut self, c: Crate) -> Crate { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Provided Methods

fn fold_item(&mut self, item: Item) -> Option<Item>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_inner_recur(&mut self, inner: ItemEnum) -> ItemEnum

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

don't override!

fn fold_item_recur(&mut self, item: Item) -> Option<Item>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

don't override!

fn fold_mod(&mut self, m: Module) -> Module

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_crate(&mut self, c: Crate) -> Crate

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?