Module rustdoc::formats::cache

Structs

Cache 🔒

This cache is used to store information about the clean::Crate being rendered in order to provide more useful documentation. This contains information like all implementors of a trait, all traits a type implements, documentation for all known traits, etc.

CacheBuilder 🔒

This struct is used to wrap the cache and tcx in order to run DocFolder.

OrphanImplItem 🔒

Enums

ParentStackItem 🔒

Information about trait and type parents is tracked while traversing the item tree to build the cache.

Functions

clean_impl_generics 🔒