Enum rustdoc::formats::item_type::ItemType

#[repr(u8)]
pub(crate) enum ItemType {

    Keyword = 0,
    Primitive = 1,
    Module = 2,
    ExternCrate = 3,
    Import = 4,
    Struct = 5,
    Enum = 6,
    Function = 7,
    TypeAlias = 8,
    Static = 9,
    Trait = 10,
    Impl = 11,
    TyMethod = 12,
    Method = 13,
    StructField = 14,
    Variant = 15,
    Macro = 16,
    AssocType = 17,
    Constant = 18,
    AssocConst = 19,
    Union = 20,
    ForeignType = 21,
    OpaqueTy = 22,
    ProcAttribute = 23,
    ProcDerive = 24,
    TraitAlias = 25,

}

Item type. Corresponds to clean::ItemEnum variants.

The search index uses item types encoded as smaller numbers which equal to discriminants. JavaScript then is used to decode them into the original value. Consequently, every change to this type should be synchronized to the itemTypes mapping table in html/static/js/search.js.

The search engine in search.js also uses item type numbers as a tie breaker when sorting results. Keywords and primitives are given first because we want them to be easily found by new users who don’t know about advanced features like type filters. The rest are mostly in an arbitrary order, but it’s easier to test the search engine when it’s deterministic, and these are strictly finer-grained than language namespaces, so using the path and the item type together to sort ensures that search sorting is stable.

In addition, code in html::render uses this enum to generate CSS classes, page prefixes, and module headings. If you are adding to this enum and want to ensure that the sidebar also prints a heading, edit the listing in html/render.rs, function sidebar_module. This uses an ordering based on a helper function inside item_module, in the same file.

Variants

Keyword = 0

Primitive = 1

Module = 2

ExternCrate = 3

Import = 4

Struct = 5

Enum = 6

Function = 7

TypeAlias = 8

Static = 9

Trait = 10

Impl = 11

TyMethod = 12

Method = 13

StructField = 14

Variant = 15

Macro = 16

AssocType = 17

Constant = 18

AssocConst = 19

Union = 20

ForeignType = 21

OpaqueTy = 22

ProcAttribute = 23

ProcDerive = 24

TraitAlias = 25

Implementations

impl ItemType

pub(crate) fn from_def_kind(kind: DefKind, parent_kind: Option<DefKind>) -> Self

Depending on the parent kind, some variants have a different translation (like a Method becoming a TyMethod).

pub(crate) fn as_str(&self) -> &'static str

pub(crate) fn is_method(&self) -> bool

pub(crate) fn is_adt(&self) -> bool