Module rustdoc::html::escape

HTML escaping.

This module contains one unit struct, which can be used to HTML-escape a string of text (for use in a format string).

Structs

Escape 🔒

Wrapper struct which will emit the HTML-escaped version of the contained string when passed to a format string.

EscapeBodyText 🔒

Wrapper struct which will emit the HTML-escaped version of the contained string when passed to a format string.