Module rustdoc::html::format

HTML formatting module

This module contains a large number of fmt::Display implementations for various types in rustdoc::clean. These implementations all currently assume that HTML output is desired, although it may be possible to redesign them in the future to instead emit any format desired.

Structs

Buffer ğŸ”’

Enums

Ending ğŸ”’

HrefError ğŸ”’

Traits

Print ğŸ”’

PrintWithSpace ğŸ”’

Functions

anchor ğŸ”’

comma_sep ğŸ”’

display_fn ğŸ”’

fmt_type ğŸ”’

generate_macro_def_id_path ğŸ”’

This function is to get the external macro path because they are not in the cache used in href_with_root_path.

href ğŸ”’

href_relative_parts ğŸ”’

Both paths should only be modules. This is because modules get their own directories; that is, std::vec and std::vec::Vec will both need ../iter/trait.Iterator.html to get at the iterator trait.

href_with_root_path ğŸ”’

join_with_double_colon ğŸ”’

primitive_link ğŸ”’

primitive_link_fragment ğŸ”’

print_abi_with_space ğŸ”’

print_constness_with_space ğŸ”’

print_default_space ğŸ”’

print_generic_bounds ğŸ”’

print_where_clause ğŸ”’

The Generics from which to emit a where-clause.

The number of spaces to indent each line with.

Whether the where-clause needs to add a comma and newline after the last bound.

resolved_path ğŸ”’

Used to render a clean::Path.

tybounds ğŸ”’

Helper to render type parameters