rustdoc::html::highlight
Struct TokenHandler

struct TokenHandler<'a, 'tcx, F: Write> {
    out: &'a mut F,
    closing_tags: Vec<(&'static str, Class)>,
    pending_exit_span: Option<Class>,
    current_class: Option<Class>,
    pending_elems: Vec<(&'a str, Option<Class>)>,
    href_context: Option<HrefContext<'a, 'tcx>>,
}

This type is used as a conveniency to prevent having to pass all its fields as arguments into the various functions (which became its methods).

Fields
out: &'a mut Fclosing_tags: Vec<(&'static str, Class)>

It contains the closing tag and the associated Class.

pending_exit_span: Option<Class>

This is used because we don’t automatically generate the closing tag on ExitSpan in case an EnterSpan event with the same class follows.

current_class: Option<Class>

current_class and pending_elems are used to group HTML elements with same class attributes to reduce the DOM size.

pending_elems: Vec<(&'a str, Option<Class>)>

We need to keep the Class for each element because it could contain a Span which is used to generate links.

href_context: Option<HrefContext<'a, 'tcx>>
Implementations

impl<F: Write> TokenHandler<'_, '_, F>

fn handle_exit_span(&mut self)

fn write_pending_elems(&mut self, current_class: Option<Class>) -> bool

Write all the pending elements sharing a same (or at mergeable) Class.

If there is a “parent” (if a EnterSpan event was encountered) and the parent can be merged with the elements’ class, then we simply write the elements since the ExitSpan event will close the tag.

Otherwise, if there is only one pending element, we let the string function handle both opening and closing the tag, otherwise we do it into this function.

It returns true if current_class must be set to None afterwards.