Enum rustdoc::html::item_type::ItemType

pub enum ItemType {
    Module,
    ExternCrate,
    Import,
    Struct,
    Enum,
    Function,
    Typedef,
    Static,
    Trait,
    Impl,
    TyMethod,
    Method,
    StructField,
    Variant,
    Macro,
    Primitive,
    AssociatedType,
    Constant,
    AssociatedConst,
}

Unstable

Item type. Corresponds to clean::ItemEnum variants.

The search index uses item types encoded as smaller numbers which equal to discriminants. JavaScript then is used to decode them into the original value. Consequently, every change to this type should be synchronized to the itemTypes mapping table in static/main.js.

Variants

Module
Unstable

ExternCrate
Unstable

Import
Unstable

Struct
Unstable

Enum
Unstable

Function
Unstable

Typedef
Unstable

Static
Unstable

Trait
Unstable

Impl
Unstable

TyMethod
Unstable

Method
Unstable

StructField
Unstable

Variant
Unstable

Macro
Unstable

Primitive
Unstable

AssociatedType
Unstable

Constant
Unstable

AssociatedConst
Unstable

Methods

impl ItemType

fn from_item(item: &Item) -> ItemType

Unstable

fn from_type_kind(kind: TypeKind) -> ItemType

Unstable

fn to_static_str(&self) -> &'static str

Unstable