Struct rustdoc::html::length_limit::HtmlWithLimit

pub(super) struct HtmlWithLimit {
    buf: String,
    len: usize,
    limit: usize,
    queued_tags: Vec<&'static str>,
    unclosed_tags: Vec<&'static str>,
}

A buffer that allows generating HTML with a length limit.

This buffer ensures that:

all tags are closed,

tags are closed in the reverse order of when they were opened (i.e., the correct HTML order),

no tags are left empty (e.g., <em></em>) due to the length limit being reached,

all text is escaped.

Fields
buf: Stringlen: usizelimit: usizequeued_tags: Vec<&'static str>

A list of tags that have been requested to be opened via Self::open_tag() but have not actually been pushed to buf yet. This ensures that tags are not left empty (e.g., <em></em>) due to the length limit being reached.

unclosed_tags: Vec<&'static str>

A list of all tags that have been opened but not yet closed.

Implementations

impl HtmlWithLimit

pub(super) fn new(length_limit: usize) -> Self

Create a new buffer, with a limit of length_limit.

pub(super) fn finish(self) -> String

Finish using the buffer and get the written output. This function will close all unclosed tags for you.

pub(super) fn push(&mut self, text: &str) -> ControlFlow<(), ()>

Write some plain text to the buffer, escaping as needed.

This function skips writing the text if the length limit was reached and returns ControlFlow::Break.

pub(super) fn open_tag(&mut self, tag_name: &'static str)

Open an HTML tag.

Note: HTML attributes have not yet been implemented. This function will panic if called with a non-alphabetic tag_name.

pub(super) fn close_tag(&mut self)

Close the most recently opened HTML tag.

fn flush_queue(&mut self)

Write all queued tags and add them to the unclosed_tags list.

fn close_all_tags(&mut self)

Close all unclosed tags.