Function rustdoc::html::markdown::reset_headers

pub fn reset_headers()

Unstable (rustdoc)

By default this markdown renderer generates anchors for each header in the rendered document. The anchor name is the contents of the header separated by hyphens, and a thread-local map is used to disambiguate among duplicate headers (numbers are appended).

This method will reset the local table for these headers. This is typically used at the beginning of rendering an entire HTML page to reset from the previous state (if any).