Struct rustdoc::html::render::context::Context

pub(crate) struct Context<'tcx> {
    pub(crate) current: Vec<Symbol>,
    pub(crate) dst: PathBuf,
    pub(super) render_redirect_pages: bool,
    pub(super) deref_id_map: FxHashMap<DefId, String>,
    pub(super) id_map: IdMap,
    pub(crate) shared: Rc<SharedContext<'tcx>>,
    pub(crate) include_sources: bool,
}

Major driving force in all rustdoc rendering. This contains information about where in the tree-like hierarchy rendering is occurring and controls how the current page is being rendered.

It is intended that this context is a lightweight object which can be fairly easily cloned because it is cloned per work-job (about once per item in the rustdoc tree).

Fields
current: Vec<Symbol>

Current hierarchy of components leading down to what’s currently being rendered

dst: PathBuf

The current destination folder of where HTML artifacts should be placed. This changes as the context descends into the module hierarchy.

render_redirect_pages: bool

A flag, which when true, will render pages which redirect to the real location of an item. This is used to allow external links to publicly reused items to redirect to the right location.

deref_id_map: FxHashMap<DefId, String>

Tracks section IDs for Deref targets so they match in both the main body and the sidebar.

id_map: IdMap

The map used to ensure all generated ‘id=’ attributes are unique.

shared: Rc<SharedContext<'tcx>>

Shared mutable state.

Issue for improving the situation: #82381

include_sources: bool

This flag indicates whether source links should be generated or not. If the source files are present in the html rendering, then this will be true.

Implementations

impl<'tcx> Context<'tcx>

pub(crate) fn tcx(&self) -> TyCtxt<'tcx>

pub(crate) fn cache(&self) -> &Cache

pub(super) fn sess(&self) -> &'tcx Session

pub(super) fn derive_id(&mut self, id: String) -> String

pub(super) fn root_path(&self) -> String

String representation of how to get back to the root path of the ‘doc/’ folder in terms of a relative URL.

fn render_item(&mut self, it: &Item, is_module: bool) -> String

fn build_sidebar_items(
&self,
m: &Module
) -> BTreeMap<String, Vec<(String, Option<String>)>>

Construct a map of items shown in the sidebar to a plain-text summary of their docs.

pub(super) fn src_href(&self, item: &Item) -> Option<String>

Generates a url appropriate for an href attribute back to the source of this item.

The url generated, when clicked, will redirect the browser back to the original source code.

If None is returned, then a source link couldn’t be generated. This may happen, for example, with externally inlined items where the source of their crate documentation isn’t known.

pub(crate) fn href_from_span(
&self,
span: Span,
with_lines: bool
) -> Option<String>

pub(crate) fn href_from_span_relative(
&self,
span: Span,
relative_to: &str
) -> Option<String>

impl Context<'_>

fn suffix_path(&self, filename: &str) -> PathBuf

fn write_shared(
&self,
resource: SharedResource<'_>,
contents: impl 'static + Send + AsRef<[u8]>,
emit: &[EmitType]
) -> Result<(), Error>

fn write_minify(
&self,
resource: SharedResource<'_>,
contents: impl 'static + Send + AsRef<str> + AsRef<[u8]>,
minify: bool,
emit: &[EmitType]
) -> Result<(), Error>