rustdoc::html::render::context
Struct SharedContext

pub(crate) struct SharedContext<'tcx> {

    pub(crate) tcx: TyCtxt<'tcx>,
    pub(crate) src_root: PathBuf,
    pub(crate) layout: Layout,
    pub(crate) local_sources: FxIndexMap<PathBuf, String>,
    pub(super) show_type_layout: bool,
    pub(super) issue_tracker_base_url: Option<String>,
    created_dirs: RefCell<FxHashSet<PathBuf>>,
    pub(super) module_sorting: ModuleSorting,
    pub(crate) style_files: Vec<StylePath>,
    pub(crate) resource_suffix: String,
    pub(crate) static_root_path: Option<String>,
    pub(crate) fs: DocFS,
    pub(super) codes: ErrorCodes,
    pub(super) playground: Option<Playground>,
    all: RefCell<AllTypes>,
    errors: Receiver<String>,
    redirections: Option<RefCell<FxHashMap<String, String>>>,
    pub(crate) span_correspondence_map: FxHashMap<Span, LinkFromSrc>,
    pub(crate) cache: Cache,
    pub(crate) call_locations: FxIndexMap<DefPathHash, FxIndexMap<PathBuf, CallData>>,
    should_merge: ShouldMerge,

}

Shared mutable state used in Context and elsewhere.

Fields
tcx: TyCtxt<'tcx>src_root: PathBuf

The path to the crate root source minus the file name. Used for simplifying paths to the highlighted source code files.

layout: Layout

This describes the layout of each page, and is not modified after creation of the context (contains info like the favicon and added html).

local_sources: FxIndexMap<PathBuf, String>

The local file sources we’ve emitted and their respective url-paths.

show_type_layout: bool

Show the memory layout of types in the docs.

issue_tracker_base_url: Option<String>

The base-URL of the issue tracker for when an item has been tagged with an issue number.

created_dirs: RefCell<FxHashSet<PathBuf>>

The directories that have already been created in this doc run. Used to reduce the number of spurious create_dir_all calls.

module_sorting: ModuleSorting

This flag indicates whether listings of modules (in the side bar and documentation itself) should be ordered alphabetically or in order of appearance (in the source code).

style_files: Vec<StylePath>

Additional CSS files to be added to the generated docs.

resource_suffix: String

Suffix to add on resource files (if suffix is “-v2” then “search-index.js” becomes “search-index-v2.js”).

static_root_path: Option<String>

Optional path string to be used to load static files on output pages. If not set, uses combinations of ../ to reach the documentation root.

fs: DocFS

The fs handle we are working with.

codes: ErrorCodesplayground: Option<Playground>all: RefCell<AllTypes>errors: Receiver<String>

Storage for the errors produced while generating documentation so they can be printed together at the end.

redirections: Option<RefCell<FxHashMap<String, String>>>

None by default, depends on the generate-redirect-map option flag. If this field is set to Some(...), it’ll store redirections and then generate a JSON file at the top level of the crate.

span_correspondence_map: FxHashMap<Span, LinkFromSrc>

Correspondence map used to link types used in the source code pages to allow to click on links to jump to the type’s definition.

cache: Cache

The Cache used during rendering.

call_locations: FxIndexMap<DefPathHash, FxIndexMap<PathBuf, CallData>>should_merge: ShouldMerge

Controls whether we read / write to cci files in the doc root. Defaults read=true, write=true

Implementations

impl SharedContext<'_>

pub(crate) fn ensure_dir(&self, dst: &Path) -> Result<(), Error>

pub(crate) fn edition(&self) -> Edition