rustdoc::html::render::print_item
Macro item_template_methods

macro_rules! item_template_methods {
    () => { ... };
    (document $($rest:tt)*) => { ... };
    (document_type_layout $($rest:tt)*) => { ... };
    (render_attributes_in_pre $($rest:tt)*) => { ... };
    (render_assoc_items $($rest:tt)*) => { ... };
    ($method:ident $($rest:tt)*) => { ... };
    ($token:tt $($rest:tt)*) => { ... };
}

Implement common methods for item template structs generated by item_template!().

NOTE: this macro is intended to be used only by item_template!().