rustdoc::html::render

Module search_index

Modules

encode ğŸ”’

Structs

EntryData ğŸ”’

FunctionData

PathData ğŸ”’

SerializedSearchIndex ğŸ”’

TypeData ğŸ”’

Enums

SerializedOptional32 ğŸ”’

SerializedOptionalString ğŸ”’

SimplifiedParam ğŸ”’

Functions

build_index ğŸ”’

Builds the search index from the collected metadata

get_fn_inputs_and_outputs ğŸ”’

Return the full list of types when bounds have been resolved.

get_function_type_for_search ğŸ”’

get_index_type ğŸ”’

get_index_type_id ğŸ”’

make_nullary_fn ğŸ”’

Create a fake nullary function.

simplify_fn_constraint ğŸ”’

simplify_fn_type ğŸ”’

The point of this function is to lower generics and types into the simplified form that the frontend search engine can use.