Struct rustdoc::html::render::SharedContext

pub struct SharedContext {
    pub src_root: PathBuf,
    pub layout: Layout,
    pub include_sources: bool,
    pub local_sources: FxHashMap<PathBuf, String>,
    pub passes: FxHashSet<String>,
    pub issue_tracker_base_url: Option<String>,
    pub css_file_extension: Option<PathBuf>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
src_root: PathBuf

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The path to the crate root source minus the file name. Used for simplifying paths to the highlighted source code files.

layout: Layout

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

This describes the layout of each page, and is not modified after creation of the context (contains info like the favicon and added html).

include_sources: bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

This flag indicates whether [src] links should be generated or not. If the source files are present in the html rendering, then this will be true.

local_sources: FxHashMap<PathBuf, String>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The local file sources we've emitted and their respective url-paths.

passes: FxHashSet<String>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

All the passes that were run on this crate.

issue_tracker_base_url: Option<String>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The base-URL of the issue tracker for when an item has been tagged with an issue number.

css_file_extension: Option<PathBuf>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The given user css file which allow to customize the generated documentation theme.