rustdoc::html

Module static_files

Static files bundled with documentation output.

All the static files are included here for centralized access in case anything other than the HTML rendering code (say, the theme checker) needs to access one of these files.

Macros

static_files 🔒

Structs

StaticFile 🔒

StaticFiles 🔒

Statics

SCRAPE_EXAMPLES_HELP_MD 🔒

STATIC_FILES 🔒

Functions

for_each 🔒

static_filename 🔒

suffix_path 🔒

Insert the provided suffix into a filename just before the extension.