Module rustdoc::html::static_files::nanum_barun_gothic

Files related to the Nanum Barun Gothic font.

These files are used to avoid some legacy CJK serif fonts in Windows.

Note that the Noto Sans KR font, which was used previously but was not very readable on Windows, has been replaced by the Nanum Barun Gothic font. This is due to Windows’ implementation of font rendering that distorts OpenType fonts too much.

The font files were generated with these commands:

pyftsubset NanumBarunGothic.ttf \
--unicodes=U+AC00-D7AF,U+1100-11FF,U+3130-318F,U+A960-A97F,U+D7B0-D7FF \
--output-file=NanumBarunGothic.ttf.woff2 --flavor=woff2

Statics

LICENSE 🔒

The file NanumBarunGothic-LICENSE.txt, the license text of the Nanum Barun Gothic font.

REGULAR 🔒

The file NanumBarunGothic.ttf.woff2, the Regular variant of the Nanum Barun Gothic font.