Module rustdoc::html::static_files::source_code_pro

Files related to the Source Code Pro font.

Statics

ITALIC 🔒

The file SourceCodePro-It.ttf.woff2, the Italic variant of the Source Code Pro font in woff2.

LICENSE 🔒

The file SourceCodePro-LICENSE.txt, the license text of the Source Code Pro font.

REGULAR 🔒

The file SourceCodePro-Regular.ttf.woff2, the Regular variant of the Source Code Pro font in woff2.

SEMIBOLD 🔒

The file SourceCodePro-Semibold.ttf.woff2, the Semibold variant of the Source Code Pro font in woff2.