Struct rustdoc::html::toc::TocBuilder

pub struct TocBuilder {
    // some fields omitted
}

Unstable (rustdoc)

Progressive construction of a table of contents.

Methods

impl TocBuilder

fn new() -> TocBuilder

fn into_toc(self) -> Toc

Unstable (rustdoc)

Convert into a true Toc struct.

fn push<'a>(&'a mut self, level: u32, name: String, id: String) -> &'a str

Unstable (rustdoc)

Push a level level heading into the appropriate place in the hierarchy, returning a string containing the section number in <num>.<num>.<num> format.