rustdoc

Module json

Rustdoc’s JSON backend

This module contains the logic for rendering a crate as JSON rather than the normal static HTML output. See the RFC and the types module docs for usage and details.

Modules

conversions 🔒

These from impls are used to create the JSON types which get serialized. They’re very close to the clean types but with some fields removed or stringified to simplify the output and not expose unstable compiler internals.

ids 🔒

Id handling for rustdoc-json.

import_finder 🔒

size_asserts 🔒

Structs

JsonRenderer 🔒

Functions

target 🔒