rustdoc::json

Module conversions

These from impls are used to create the JSON types which get serialized. They’re very close to the clean types but with some fields removed or stringified to simplify the output and not expose unstable compiler internals.

Traits

FromClean 🔒

IntoJson 🔒

Functions

format_integer_type 🔒

from_clean_function 🔒

from_clean_item 🔒

from_clean_static 🔒

maybe_from_hir_attr 🔒

Maybe convert a attribute from hir to json.

other_attr 🔒

repr_attr 🔒