rustdoc::json
Struct JsonRenderer

pub(crate) struct JsonRenderer<'tcx> {
    tcx: TyCtxt<'tcx>,
    index: Rc<RefCell<FxHashMap<Id, Item>>>,
    out_dir: Option<PathBuf>,
    cache: Rc<Cache>,
    imported_items: DefIdSet,
    id_interner: Rc<RefCell<FxHashMap<(FullItemId, Option<FullItemId>), Id>>>,
}

Fields
tcx: TyCtxt<'tcx>index: Rc<RefCell<FxHashMap<Id, Item>>>

A mapping of IDs that contains all local items for this crate which gets output as a top level field of the JSON blob.

out_dir: Option<PathBuf>

The directory where the JSON blob should be written to.

If this is None, the blob will be printed to stdout instead.

cache: Rc<Cache>imported_items: DefIdSetid_interner: Rc<RefCell<FxHashMap<(FullItemId, Option<FullItemId>), Id>>>
Implementations

impl JsonRenderer<'_>

pub(super) fn convert_item(&self, item: Item) -> Option<Item>

fn convert_span(&self, span: Span) -> Option<Span>

fn convert_visibility(&self, v: Option<Visibility<DefId>>) -> Visibility

pub(crate) fn id_from_item_default(&self, item_id: ItemId) -> Id

pub(crate) fn id_from_item_inner( &self, item_id: ItemId, name: Option<Symbol>, extra: Option<Id>, ) -> Id

pub(crate) fn id_from_item(&self, item: &Item) -> Id

fn ids(&self, items: impl IntoIterator<Item = Item>) -> Vec<Id>

fn ids_keeping_stripped( &self, items: impl IntoIterator<Item = Item>, ) -> Vec<Option<Id>>

impl<'tcx> JsonRenderer<'tcx>

fn sess(&self) -> &'tcx Session

fn get_trait_implementors(&mut self, id: DefId) -> Vec<Id>

fn get_impls(&mut self, id: DefId) -> Vec<Id>

fn serialize_and_write<T: Write>( &self, output_crate: Crate, writer: BufWriter<T>, path: &str, ) -> Result<(), Error>