Module rustdoc::passes::collect_trait_impls

Collects trait impls for each item in the crate. For example, if a crate defines a struct that implements a trait, this pass will note that the struct implements that trait.

Structs

BadImplStripper 🔒

ItemAndAliasCollector 🔒

SyntheticImplCollector 🔒

Constants

COLLECT_TRAIT_IMPLS 🔒

Functions

collect_trait_impls 🔒