Struct rustdoc::plugins::PluginManager

pub struct PluginManager {
    pub prefix: PathBuf,
    // some fields omitted
}

Unstable (rustdoc)

Manages loading and running of plugins

Fields
prefix: PathBuf
Unstable (rustdoc)

The directory plugins will be loaded from

Methods

impl PluginManager

fn new(prefix: PathBuf) -> PluginManager

Unstable (rustdoc)

Create a new plugin manager

fn load_plugin(&mut self, name: String)

Unstable (rustdoc)

Load a plugin with the given name.

Turns name into the proper dynamic library filename for the given platform. On windows, it turns into name.dll, on OS X, name.dylib, and elsewhere, libname.so.

fn add_plugin(&mut self, plugin: PluginCallback)

Unstable (rustdoc)

Load a normal Rust function as a plugin.

This is to run passes over the cleaned crate. Plugins run this way correspond to the A-aux tag on Github.

fn run_plugins(&self, krate: Crate) -> Crate

Unstable (rustdoc)

Run all the loaded plugins over the crate, returning their results