rustdoc
Module visit_ast

The Rust AST Visitor. Extracts useful information and massages it into a form usable for clean.

Structs

Module 🔒

This module is used to store stuff from Rust’s AST in a more convenient manner (and with prettier names) before cleaning.

RustdocVisitor 🔒

Functions

def_id_to_path 🔒