Struct rustdoc::visit_ast::RustdocVisitor

pub(crate) struct RustdocVisitor<'a, 'tcx> {
    cx: &'a mut DocContext<'tcx>,
    view_item_stack: FxHashSet<HirId>,
    inlining: bool,
    inside_public_path: bool,
    exact_paths: FxHashMap<DefId, Vec<Symbol>>,
}

Fields
cx: &'a mut DocContext<'tcx>view_item_stack: FxHashSet<HirId>inlining: boolinside_public_path: bool

Are the current module and all of its parents public?

exact_paths: FxHashMap<DefId, Vec<Symbol>>
Implementations

impl<'a, 'tcx> RustdocVisitor<'a, 'tcx>

pub(crate) fn new(cx: &'a mut DocContext<'tcx>) -> RustdocVisitor<'a, 'tcx>

fn store_path(&mut self, did: DefId)

pub(crate) fn visit(self) -> Module<'tcx>

fn visit_mod_contents(
&mut self,
id: HirId,
m: &'tcx Mod<'tcx>,
name: Symbol
) -> Module<'tcx>

fn maybe_inline_local(
&mut self,
id: HirId,
res: Res,
renamed: Option<Symbol>,
glob: bool,
om: &mut Module<'tcx>,
please_inline: bool
) -> bool

Tries to resolve the target of a pub use statement and inlines the target if it is defined locally and would not be documented otherwise, or when it is specifically requested with please_inline. (the latter is the case when the import is marked doc(inline))

Cross-crate inlining occurs later on during crate cleaning and follows different rules.

Returns true if the target has been inlined.

fn visit_item(
&mut self,
item: &'tcx Item<'_>,
renamed: Option<Symbol>,
om: &mut Module<'tcx>
)

fn visit_foreign_item(
&mut self,
item: &'tcx ForeignItem<'_>,
renamed: Option<Symbol>,
om: &mut Module<'tcx>
)