Crate rustdocUnstable

Reexports

pub use clean::SCHEMA_VERSION;

Modules

clean
This module contains the "cleaned" pieces of the AST, and the functions that clean them.

core

doctree
This module is used to store stuff from Rust's AST in a more convenient manner (and with prettier names) before cleaning.

externalfiles

fold

html

markdown

passes

plugins

stability_summary
This module crawls a clean::Crate and produces a summarization of the stability levels within the crate. The summary contains the module hierarchy, with item counts for every stability level per module. A parent module's count includes its children's.

test

visit_ast
Rust AST Visitor. Extracts useful information and massages it into a form usable for clean

Statics

ANALYSISKEY

Functions

main

main_args

opts

usage