Enum rustdoc::clean::cfg::Cfg

pub(crate) enum Cfg {
    True,
    False,
    Cfg(Symbol, Option<Symbol>),
    Not(Box<Cfg>),
    Any(Vec<Cfg>),
    All(Vec<Cfg>),
}

Variants

True

Accepts all configurations.

False

Denies all configurations.

Cfg(Symbol, Option<Symbol>)

A generic configuration option, e.g., test or target_os = "linux".

Not(Box<Cfg>)

Negates a configuration requirement, i.e., not(x).

Any(Vec<Cfg>)

Union of a list of configuration requirements, i.e., any(...).

All(Vec<Cfg>)

Intersection of a list of configuration requirements, i.e., all(...).

Implementations

impl Cfg

fn parse_nested(
nested_cfg: &NestedMetaItem,
exclude: &FxHashSet<Cfg>
) -> Result<Option<Cfg>, InvalidCfgError>

Parses a NestedMetaItem into a Cfg.

pub(crate) fn parse_without(
cfg: &MetaItem,
exclude: &FxHashSet<Cfg>
) -> Result<Option<Cfg>, InvalidCfgError>

pub(crate) fn parse(cfg: &MetaItem) -> Result<Cfg, InvalidCfgError>

Parses a MetaItem into a Cfg.

The MetaItem should be the content of the #[cfg(...)], e.g., unix or target_os = "redox".

If the content is not properly formatted, it will return an error indicating what and where the error is.

pub(crate) fn matches(
&self,
parse_sess: &ParseSess,
features: Option<&Features>
) -> bool

Checks whether the given configuration can be matched in the current session.

Equivalent to attr::cfg_matches.

fn is_simple(&self) -> bool

Whether the configuration consists of just Cfg or Not.

fn is_all(&self) -> bool

Whether the configuration consists of just Cfg, Not or All.

pub(crate) fn render_short_html(&self) -> String

Renders the configuration for human display, as a short HTML description.

pub(crate) fn render_long_html(&self) -> String

Renders the configuration for long display, as a long HTML description.

pub(crate) fn render_long_plain(&self) -> String

Renders the configuration for long display, as a long plain text description.

fn should_capitalize_first_letter(&self) -> bool

fn should_append_only_to_description(&self) -> bool

fn should_use_with_in_description(&self) -> bool

pub(crate) fn simplify_with(&self, assume: &Cfg) -> Option<Cfg>

Attempt to simplify this cfg by assuming that assume is already known to be true, will return None if simplification managed to completely eliminate any requirements from this Cfg.

See tests::test_simplify_with for examples.