Enum rustdoc::clean::Type

pub enum Type {
    ResolvedPath {
        path: Path,
        typarams: Option<Vec<TyParamBound>>,
        did: DefId,
        is_generic: bool,
    },
    Generic(String),
    Primitive(PrimitiveType),
    BareFunction(Box<BareFunctionDecl>),
    Tuple(Vec<Type>),
    Vector(Box<Type>),
    FixedVector(Box<Type>, String),
    Never,
    Unique(Box<Type>),
    RawPointer(Mutability, Box<Type>),
    BorrowedRef {
        lifetime: Option<Lifetime>,
        mutability: Mutability,
        type_: Box<Type>,
    },
    QPath {
        name: String,
        self_type: Box<Type>,
        trait_: Box<Type>,
    },
    Infer,
    ImplTrait(Vec<TyParamBound>),
}

🔬 This is a nightly-only experimental API. (rustdoc)

A representation of a Type suitable for hyperlinking purposes. Ideally one can get the original type out of the AST/TyCtxt given one of these, if more information is needed. Most importantly it does not preserve mutability or boxes.

Variants
ResolvedPath

🔬 This is a nightly-only experimental API. (rustdoc)

structs/enums/traits (most that'd be an hir::TyPath)

Fields of ResolvedPath

path: Path

🔬 This is a nightly-only experimental API. (rustdoc)

typarams: Option<Vec<TyParamBound>>

🔬 This is a nightly-only experimental API. (rustdoc)

did: DefId

🔬 This is a nightly-only experimental API. (rustdoc)

is_generic: bool

🔬 This is a nightly-only experimental API. (rustdoc)

true if is a T::Name path for associated types

Generic(String)

🔬 This is a nightly-only experimental API. (rustdoc)

For parameterized types, so the consumer of the JSON don't go looking for types which don't exist anywhere.

Primitive(PrimitiveType)

🔬 This is a nightly-only experimental API. (rustdoc)

Primitives are the fixed-size numeric types (plus int/usize/float), char, arrays, slices, and tuples.

BareFunction(Box<BareFunctionDecl>)

🔬 This is a nightly-only experimental API. (rustdoc)

extern "ABI" fn

Tuple(Vec<Type>)

🔬 This is a nightly-only experimental API. (rustdoc)

Vector(Box<Type>)

🔬 This is a nightly-only experimental API. (rustdoc)

FixedVector(Box<Type>, String)

🔬 This is a nightly-only experimental API. (rustdoc)

Never

🔬 This is a nightly-only experimental API. (rustdoc)

Unique(Box<Type>)

🔬 This is a nightly-only experimental API. (rustdoc)

RawPointer(Mutability, Box<Type>)

🔬 This is a nightly-only experimental API. (rustdoc)

BorrowedRef

🔬 This is a nightly-only experimental API. (rustdoc)

Fields of BorrowedRef

lifetime: Option<Lifetime>

🔬 This is a nightly-only experimental API. (rustdoc)

mutability: Mutability

🔬 This is a nightly-only experimental API. (rustdoc)

type_: Box<Type>

🔬 This is a nightly-only experimental API. (rustdoc)

QPath

🔬 This is a nightly-only experimental API. (rustdoc)

Fields of QPath

name: String

🔬 This is a nightly-only experimental API. (rustdoc)

self_type: Box<Type>

🔬 This is a nightly-only experimental API. (rustdoc)

trait_: Box<Type>

🔬 This is a nightly-only experimental API. (rustdoc)

Infer

🔬 This is a nightly-only experimental API. (rustdoc)

ImplTrait(Vec<TyParamBound>)

🔬 This is a nightly-only experimental API. (rustdoc)

Methods

impl Type

fn primitive_type(&self) -> Option<PrimitiveType>

🔬 This is a nightly-only experimental API. (rustdoc)

fn is_generic(&self) -> bool

🔬 This is a nightly-only experimental API. (rustdoc)

fn is_self_type(&self) -> bool

🔬 This is a nightly-only experimental API. (rustdoc)