rustdoc::clean

Module inline

Support for inlining external documentation into the current AST.

Functions

build_const_item ğŸ”’

build_enum ğŸ”’

build_function ğŸ”’

build_impl ğŸ”’

Inline an impl, inherent or of a trait. The did must be for an impl.

build_impls ğŸ”’

Builds all inherent implementations of an ADT (struct/union/enum) or Trait item/path/reexport.

build_macro ğŸ”’

build_module ğŸ”’

build_module_items ğŸ”’

build_static ğŸ”’

build_struct ğŸ”’

build_trait ğŸ”’

build_trait_alias ğŸ”’

build_type_alias ğŸ”’

build_union ğŸ”’

item_relative_path ğŸ”’

load_attrs ğŸ”’

merge_attrs ğŸ”’

print_inlined_const ğŸ”’

record_extern_fqn ğŸ”’

Record an external fully qualified name in the external_paths cache.

record_extern_trait ğŸ”’

separate_self_bounds ğŸ”’

try_inline ğŸ”’

Attempt to inline a definition into this AST.

try_inline_glob ğŸ”’