Function rustdoc::clean::inline::try_inline

pub fn try_inline(cx: &DocContext, id: NodeId, into: Option<Name>) -> Option<Vec<Item>>

Unstable (rustdoc)

Attempt to inline the definition of a local node id into this AST.

This function will fetch the definition of the id specified, and if it is from another crate it will attempt to inline the documentation from the other crate into this crate.

This is primarily used for pub use statements which are, in general, implementation details. Inlining the documentation should help provide a better experience when reading the documentation in this use case.

The returned value is None if the id could not be inlined, and Some of a vector of items if it was successfully expanded.