rustdoc::clean::types
Enum ItemKind

pub(crate) enum ItemKind {

    ExternCrateItem {
        src: Option<Symbol>,
    },
    ImportItem(Import),
    StructItem(Struct),
    UnionItem(Union),
    EnumItem(Enum),
    FunctionItem(Box<Function>),
    ModuleItem(Module),
    TypeAliasItem(Box<TypeAlias>),
    StaticItem(Static),
    TraitItem(Box<Trait>),
    TraitAliasItem(TraitAlias),
    ImplItem(Box<Impl>),
    TyMethodItem(Box<Function>),
    MethodItem(Box<Function>, Option<Defaultness>),
    StructFieldItem(Type),
    VariantItem(Variant),
    ForeignFunctionItem(Box<Function>, Safety),
    ForeignStaticItem(Static, Safety),
    ForeignTypeItem,
    MacroItem(Macro),
    ProcMacroItem(ProcMacro),
    PrimitiveItem(PrimitiveType),
    TyAssocConstItem(Generics, Box<Type>),
    ConstantItem(Box<Constant>),
    AssocConstItem(Box<Constant>),
    TyAssocTypeItem(Generics, Vec<GenericBound>),
    AssocTypeItem(Box<TypeAlias>, Vec<GenericBound>),
    StrippedItem(Box<ItemKind>),
    KeywordItem,

}

Variants

ExternCrateItem

Fields

src: Option<Symbol>

The crateโs name, not the name itโs imported as.

ImportItem(Import)

StructItem(Struct)

UnionItem(Union)

EnumItem(Enum)

FunctionItem(Box<Function>)

ModuleItem(Module)

TypeAliasItem(Box<TypeAlias>)

StaticItem(Static)

TraitItem(Box<Trait>)

TraitAliasItem(TraitAlias)

ImplItem(Box<Impl>)

TyMethodItem(Box<Function>)

A required method in a trait declaration meaning itโs only a function signature.

MethodItem(Box<Function>, Option<Defaultness>)

A method in a trait impl or a provided method in a trait declaration.

Compared to TyMethodItem, it also contains a method body.

StructFieldItem(Type)

VariantItem(Variant)

ForeignFunctionItem(Box<Function>, Safety)

fns from an extern block

ForeignStaticItem(Static, Safety)

statics from an extern block

ForeignTypeItem

types from an extern block

MacroItem(Macro)

ProcMacroItem(ProcMacro)

PrimitiveItem(PrimitiveType)

TyAssocConstItem(Generics, Box<Type>)

A required associated constant in a trait declaration.

ConstantItem(Box<Constant>)

AssocConstItem(Box<Constant>)

An associated constant in a trait impl or a provided one in a trait declaration.

TyAssocTypeItem(Generics, Vec<GenericBound>)

A required associated type in a trait declaration.

The bounds may be non-empty if there is a where clause.

AssocTypeItem(Box<TypeAlias>, Vec<GenericBound>)

An associated type in a trait impl or a provided one in a trait declaration.

StrippedItem(Box<ItemKind>)

An item that has been stripped by a rustdoc pass

KeywordItem

Implementations

impl ItemKind

pub(crate) fn inner_items(&self) -> impl Iterator<Item = &Item>

Some items contain others such as structs (for their fields) and Enums (for their variants). This method returns those contained items.

pub(crate) fn is_non_assoc(&self) -> bool

Returns true if this item does not appear inside an impl block.