rustdoc::clean::types
Enum PrimitiveType

pub(crate) enum PrimitiveType {

    Isize,
    I8,
    I16,
    I32,
    I64,
    I128,
    Usize,
    U8,
    U16,
    U32,
    U64,
    U128,
    F16,
    F32,
    F64,
    F128,
    Char,
    Bool,
    Str,
    Slice,
    Array,
    Pat,
    Tuple,
    Unit,
    RawPointer,
    Reference,
    Fn,
    Never,

}

A primitive (aka, builtin) type.

This represents things like i32, str, etc.

N.B. This has to be different from hir::PrimTy because it also includes types that arenโt paths, like Self::Unit.

Variants

Isize

I8

I16

I32

I64

I128

Usize

U8

U16

U32

U64

U128

F16

F32

F64

F128

Char

Bool

Str

Slice

Array

Pat

Tuple

Unit

RawPointer

Reference

Fn

Never

Implementations

impl PrimitiveType

pub(crate) fn from_hir(prim: PrimTy) -> PrimitiveType

pub(crate) fn from_symbol(s: Symbol) -> Option<PrimitiveType>

pub(crate) fn simplified_types() -> &'static FxIndexMap<PrimitiveType, ArrayVec<SimplifiedType, 3>>

pub(crate) fn impls<'tcx>( &self, tcx: TyCtxt<'tcx>, ) -> impl Iterator<Item = DefId> + 'tcx

pub(crate) fn all_impls(tcx: TyCtxt<'_>) -> impl Iterator<Item = DefId> + '_

pub(crate) fn as_sym(&self) -> Symbol

pub(crate) fn primitive_locations( tcx: TyCtxt<'_>, ) -> &FxIndexMap<PrimitiveType, DefId>

Returns the DefId of the module with rustc_doc_primitive for this primitive type. Panics if there is no such module.

This gives precedence to primitives defined in the current crate, and deprioritizes primitives defined in core, but otherwise, if multiple crates define the same primitive, there is no guarantee of which will be picked.

In particular, if a crate depends on both std and another crate that also defines rustc_doc_primitive, then itโs entirely random whether std or the other crate is picked. (no_std crates are usually fine unless multiple dependencies define a primitive.)