rustdoc::clean::types
Struct Item

pub(crate) struct Item {
    pub(crate) name: Option<Symbol>,
    pub(crate) inner: Box<ItemInner>,
    pub(crate) item_id: ItemId,
    pub(crate) inline_stmt_id: Option<LocalDefId>,
    pub(crate) cfg: Option<Arc<Cfg>>,
}

Anything with a source location and set of attributes and, optionally, a name. That is, anything that can be documented. This doesn’t correspond directly to the AST’s concept of an item; it’s a strict superset.

Fields
name: Option<Symbol>

The name of this item. Optional because not every item has a name, e.g. impls.

inner: Box<ItemInner>item_id: ItemIdinline_stmt_id: Option<LocalDefId>

This is the LocalDefId of the use statement if the item was inlined. The crate metadata doesn’t hold this information, so the use statement always belongs to the current crate.

cfg: Option<Arc<Cfg>>
Implementations

impl Item

pub(crate) fn stability(&self, tcx: TyCtxt<'_>) -> Option<Stability>

Returns the effective stability of the item.

This method should only be called after the propagate-stability pass has been run.

pub(crate) fn const_stability(&self, tcx: TyCtxt<'_>) -> Option<ConstStability>

pub(crate) fn deprecation(&self, tcx: TyCtxt<'_>) -> Option<Deprecation>

pub(crate) fn inner_docs(&self, tcx: TyCtxt<'_>) -> bool

pub(crate) fn span(&self, tcx: TyCtxt<'_>) -> Option<Span>

pub(crate) fn attr_span(&self, tcx: TyCtxt<'_>) -> Span

pub(crate) fn doc_value(&self) -> String

Combine all doc strings into a single value handling indentation and newlines as needed.

pub(crate) fn opt_doc_value(&self) -> Option<String>

Combine all doc strings into a single value handling indentation and newlines as needed. Returns None is there’s no documentation at all, and Some("") if there is some documentation but it is empty (e.g. #[doc = ""]).

pub(crate) fn from_def_id_and_parts( def_id: DefId, name: Option<Symbol>, kind: ItemKind, cx: &mut DocContext<'_>, ) -> Item

pub(crate) fn from_def_id_and_attrs_and_parts( def_id: DefId, name: Option<Symbol>, kind: ItemKind, attrs: Attributes, cfg: Option<Arc<Cfg>>, ) -> Item

pub(crate) fn links(&self, cx: &Context<'_>) -> Vec<RenderedLink>

pub(crate) fn link_names(&self, cache: &Cache) -> Vec<RenderedLink>

Find a list of all link names, without finding their href.

This is used for generating summary text, which does not include the link text, but does need to know which []-bracketed names are actually links.

pub(crate) fn is_crate(&self) -> bool

pub(crate) fn is_mod(&self) -> bool

pub(crate) fn is_struct(&self) -> bool

pub(crate) fn is_enum(&self) -> bool

pub(crate) fn is_variant(&self) -> bool

pub(crate) fn is_associated_type(&self) -> bool

pub(crate) fn is_ty_associated_type(&self) -> bool

pub(crate) fn is_associated_const(&self) -> bool

pub(crate) fn is_ty_associated_const(&self) -> bool

pub(crate) fn is_method(&self) -> bool

pub(crate) fn is_ty_method(&self) -> bool

pub(crate) fn is_primitive(&self) -> bool

pub(crate) fn is_union(&self) -> bool

pub(crate) fn is_import(&self) -> bool

pub(crate) fn is_extern_crate(&self) -> bool

pub(crate) fn is_keyword(&self) -> bool

pub(crate) fn is_stripped(&self) -> bool

pub(crate) fn has_stripped_entries(&self) -> Option<bool>

pub(crate) fn stability_class(&self, tcx: TyCtxt<'_>) -> Option<String>

pub(crate) fn stable_since(&self, tcx: TyCtxt<'_>) -> Option<StableSince>

pub(crate) fn is_non_exhaustive(&self) -> bool

pub(crate) fn type_(&self) -> ItemType

Returns a documentation-level item type from the item.

pub(crate) fn is_default(&self) -> bool

pub(crate) fn fn_header(&self, tcx: TyCtxt<'_>) -> Option<FnHeader>

Returns a FnHeader if self is a function item, otherwise returns None.

pub(crate) fn visibility(&self, tcx: TyCtxt<'_>) -> Option<Visibility<DefId>>

Returns the visibility of the current item. If the visibility is “inherited”, then None is returned.

pub(crate) fn attributes( &self, tcx: TyCtxt<'_>, cache: &Cache, keep_as_is: bool, ) -> Vec<String>

pub fn is_doc_hidden(&self) -> bool

pub fn def_id(&self) -> Option<DefId>