Module rustdoc::clean::utils

Constants

DOC_RUST_LANG_ORG_CHANNEL ğŸ”’

A link to doc.rust-lang.org that includes the channel name. Use this instead of manual links so that the channel is consistent.

Statics

DOC_CHANNEL ğŸ”’

Functions

build_deref_target_impls ğŸ”’

display_macro_source ğŸ”’

enter_impl_trait ğŸ”’

external_generic_args ğŸ”’

external_path ğŸ”’

find_nearest_parent_module ğŸ”’

Find the nearest parent module of a DefId.

format_integer_with_underscore_sep ğŸ”’

get_auto_trait_and_blanket_impls ğŸ”’

has_doc_flag ğŸ”’

Checks for the existence of hidden in the attribute below if flag is sym::hidden:

inherits_doc_hidden ğŸ”’

is_literal_expr ğŸ”’

krate ğŸ”’

name_from_pat ğŸ”’

print_const ğŸ”’

print_const_expr ğŸ”’

Build a textual representation of an unevaluated constant expression.

print_const_with_custom_print_scalar ğŸ”’

print_evaluated_const ğŸ”’

qpath_to_string ğŸ”’

register_res ğŸ”’

If res has a documentation page associated, store it in the cache.

render_macro_arms ğŸ”’

Render a sequence of macro arms in a format suitable for displaying to the user as part of an item declaration.

resolve_type ğŸ”’

Given a type Path, resolve it to a Type using the TyCtxt

resolve_use_source ğŸ”’

should_ignore_res ğŸ”’

strip_path_generics ğŸ”’

Remove the generic arguments from a path.

ty_args_to_args ğŸ”’