rustdoc::config
Struct Options

pub(crate) struct Options {

    pub(crate) crate_name: Option<String>,
    pub(crate) bin_crate: bool,
    pub(crate) proc_macro_crate: bool,
    pub(crate) error_format: ErrorOutputType,
    pub(crate) diagnostic_width: Option<usize>,
    pub(crate) libs: Vec<SearchPath>,
    pub(crate) lib_strs: Vec<String>,
    pub(crate) externs: Externs,
    pub(crate) extern_strs: Vec<String>,
    pub(crate) cfgs: Vec<String>,
    pub(crate) check_cfgs: Vec<String>,
    pub(crate) codegen_options: CodegenOptions,
    pub(crate) codegen_options_strs: Vec<String>,
    pub(crate) unstable_opts: UnstableOptions,
    pub(crate) unstable_opts_strs: Vec<String>,
    pub(crate) target: TargetTuple,
    pub(crate) edition: Edition,
    pub(crate) maybe_sysroot: Option<PathBuf>,
    pub(crate) lint_opts: Vec<(String, Level)>,
    pub(crate) describe_lints: bool,
    pub(crate) lint_cap: Option<Level>,
    pub(crate) should_test: bool,
    pub(crate) test_args: Vec<String>,
    pub(crate) test_run_directory: Option<PathBuf>,
    pub(crate) persist_doctests: Option<PathBuf>,
    pub(crate) runtool: Option<String>,
    pub(crate) runtool_args: Vec<String>,
    pub(crate) enable_per_target_ignores: bool,
    pub(crate) no_run: bool,
    pub(crate) remap_path_prefix: Vec<(PathBuf, PathBuf)>,
    pub(crate) test_builder: Option<PathBuf>,
    pub(crate) test_builder_wrappers: Vec<PathBuf>,
    pub(crate) show_coverage: bool,
    pub(crate) crate_version: Option<String>,
    pub(crate) output_format: OutputFormat,
    pub(crate) run_check: bool,
    pub(crate) json_unused_externs: JsonUnusedExterns,
    pub(crate) nocapture: bool,
    pub(crate) scrape_examples_options: Option<ScrapeExamplesOptions>,
    pub(crate) unstable_features: UnstableFeatures,
    pub(crate) expanded_args: Vec<String>,

}

Configuration options for rustdoc.

Fields
crate_name: Option<String>

The name of the crate being documented.

bin_crate: bool

Whether or not this is a bin crate

proc_macro_crate: bool

Whether or not this is a proc-macro crate

error_format: ErrorOutputType

How to format errors and warnings.

diagnostic_width: Option<usize>

Width of output buffer to truncate errors appropriately.

libs: Vec<SearchPath>

Library search paths to hand to the compiler.

lib_strs: Vec<String>

Library search paths strings to hand to the compiler.

externs: Externs

The list of external crates to link against.

extern_strs: Vec<String>

The list of external crates strings to link against.

cfgs: Vec<String>

List of cfg flags to hand to the compiler. Always includes rustdoc.

check_cfgs: Vec<String>

List of check cfg flags to hand to the compiler.

codegen_options: CodegenOptions

Codegen options to hand to the compiler.

codegen_options_strs: Vec<String>

Codegen options strings to hand to the compiler.

unstable_opts: UnstableOptions

Unstable (-Z) options to pass to the compiler.

unstable_opts_strs: Vec<String>

Unstable (-Z) options strings to pass to the compiler.

target: TargetTuple

The target used to compile the crate against.

edition: Edition

Edition used when reading the crate. Defaults to “2015”. Also used by default when compiling doctests from the crate.

maybe_sysroot: Option<PathBuf>

The path to the sysroot. Used during the compilation process.

lint_opts: Vec<(String, Level)>

Lint information passed over the command-line.

describe_lints: bool

Whether to ask rustc to describe the lints it knows.

lint_cap: Option<Level>

What level to cap lints at.

should_test: bool

Whether we should run doctests instead of generating docs.

test_args: Vec<String>

List of arguments to pass to the test harness, if running tests.

test_run_directory: Option<PathBuf>

The working directory in which to run tests.

persist_doctests: Option<PathBuf>

Optional path to persist the doctest executables to, defaults to a temporary directory if not set.

runtool: Option<String>

Runtool to run doctests with

runtool_args: Vec<String>

Arguments to pass to the runtool

enable_per_target_ignores: bool

Whether to allow ignoring doctests on a per-target basis For example, using ignore-foo to ignore running the doctest on any target that contains “foo” as a substring

no_run: bool

Do not run doctests, compile them if should_test is active.

remap_path_prefix: Vec<(PathBuf, PathBuf)>

What sources are being mapped.

test_builder: Option<PathBuf>

The path to a rustc-like binary to build tests with. If not set, we default to loading from $sysroot/bin/rustc.

test_builder_wrappers: Vec<PathBuf>

Run these wrapper instead of rustc directly

show_coverage: bool

Whether to run the calculate-doc-coverage pass, which counts the number of public items with and without documentation.

crate_version: Option<String>

Crate version to note on the sidebar of generated docs.

output_format: OutputFormat

The format that we output when rendering.

Currently used only for the --show-coverage option.

run_check: bool

If this option is set to true, rustdoc will only run checks and not generate documentation.

json_unused_externs: JsonUnusedExterns

Whether doctests should emit unused externs

nocapture: bool

Whether to skip capturing stdout and stderr of tests.

scrape_examples_options: Option<ScrapeExamplesOptions>

Configuration for scraping examples from the current crate. If this option is Some(..) then the compiler will scrape examples and not generate documentation.

unstable_features: UnstableFeatures

Note: this field is duplicated in RenderOptions because it’s useful to have it in both places.

expanded_args: Vec<String>

All commandline args used to invoke the compiler, with @file args fully expanded. This will only be used within debug info, e.g. in the pdb file on windows This is mainly useful for other tools that reads that debuginfo to figure out how to call the compiler with the same arguments.

Implementations

impl Options

pub(crate) fn from_matches( early_dcx: &mut EarlyDiagCtxt, matches: &Matches, args: Vec<String>, ) -> Option<(InputMode, Options, RenderOptions)>

Parses the given command-line for options. If an error message or other early-return has been printed, returns Err with the exit code.