rustdoc

Module core

Structs

DocContext 🔒

EmitIgnoredResolutionErrors 🔒

Due to https://github.com/rust-lang/rust/pull/73566, the name resolution pass may find errors that are never emitted. If typeck is called after this happens, then we’ll get an ICE: ‘Res::Error found but not reported’. To avoid this, emit the errors now.

Options 🔒

The top-level command-line options struct.

UnstableOptions 🔒

Enums

ImplTraitParam 🔒

DefId or parameter index (ty::ParamTy.index) of a synthetic type parameter for impl Trait in argument position.

Functions

create_config 🔒

Parse, resolve, and typecheck the given crate.

new_dcx 🔒

Creates a new DiagCtxt that can be used to emit warnings and errors.

run_global_ctxt 🔒