Struct rustdoc::core::DocContext

pub(crate) struct DocContext<'tcx> {

    pub(crate) tcx: TyCtxt<'tcx>,
    pub(crate) resolver: Rc<RefCell<BoxedResolver>>,
    pub(crate) resolver_caches: ResolverCaches,
    pub(crate) param_env: ParamEnv<'tcx>,
    pub(crate) external_traits: Rc<RefCell<FxHashMap<DefId, TraitWithExtraInfo>>>,
    pub(crate) active_extern_traits: FxHashSet<DefId>,
    pub(crate) substs: FxHashMap<DefId, SubstParam>,
    pub(crate) impl_trait_bounds: FxHashMap<ImplTraitParam, Vec<GenericBound>>,
    pub(crate) generated_synthetics: FxHashSet<(Ty<'tcx>, DefId)>,
    pub(crate) auto_traits: Vec<DefId>,
    pub(crate) render_options: RenderOptions,
    pub(crate) cache: Cache,
    pub(crate) inlined: FxHashSet<ItemId>,
    pub(crate) output_format: OutputFormat,
    pub(crate) show_coverage: bool,

}

Fields
tcx: TyCtxt<'tcx>resolver: Rc<RefCell<BoxedResolver>>

Name resolver. Used for intra-doc links.

The Rc<RefCell<...>> wrapping is needed because that is what’s returned by rustc_interface::Queries::expansion().

resolver_caches: ResolverCachesparam_env: ParamEnv<'tcx>

Used for normalization.

Most of this logic is copied from rustc_lint::late.

external_traits: Rc<RefCell<FxHashMap<DefId, TraitWithExtraInfo>>>

Later on moved through clean::Crate into cache

active_extern_traits: FxHashSet<DefId>

Used while populating external_traits to ensure we don’t process the same trait twice at the same time.

substs: FxHashMap<DefId, SubstParam>

Table DefId of type, lifetime, or const parameter -> substituted type, lifetime, or const

impl_trait_bounds: FxHashMap<ImplTraitParam, Vec<GenericBound>>

Table synthetic type parameter for impl Trait in argument position -> bounds

generated_synthetics: FxHashSet<(Ty<'tcx>, DefId)>

Auto-trait or blanket impls processed so far, as (self_ty, trait_def_id).

auto_traits: Vec<DefId>render_options: RenderOptions

The options given to rustdoc that could be relevant to a pass.

cache: Cache

This same cache is used throughout rustdoc, including in crate::html::render.

inlined: FxHashSet<ItemId>

Used by clean::inline to tell if an item has already been inlined.

output_format: OutputFormat

Used by calculate_doc_coverage.

show_coverage: bool

Used by strip_private.

Implementations

impl<'tcx> DocContext<'tcx>

pub(crate) fn sess(&self) -> &'tcx Session

pub(crate) fn with_param_env<T, F: FnOnce(&mut Self) -> T>(
&mut self,
def_id: DefId,
f: F
) -> T

pub(crate) fn enter_resolver<F, R>(&self, f: F) -> Rwhere
F: FnOnce(&mut Resolver<'_>) -> R,

pub(crate) fn enter_alias<F, R>(
&mut self,
substs: FxHashMap<DefId, SubstParam>,
f: F
) -> Rwhere
F: FnOnce(&mut Self) -> R,

Call the closure with the given parameters set as the substitutions for a type alias’ RHS.

pub(crate) fn as_local_hir_id(tcx: TyCtxt<'_>, item_id: ItemId) -> Option<HirId>

Like hir().local_def_id_to_hir_id(), but skips calling it on fake DefIds. (This avoids a slice-index-out-of-bounds panic.)

pub(crate) fn with_all_traits(&mut self, f: impl FnOnce(&mut Self, &[DefId]))

pub(crate) fn with_all_trait_impls(
&mut self,
f: impl FnOnce(&mut Self, &[DefId])
)