rustdoc::html
Module format

HTML formatting module

This module contains a large number of Display implementations for various types in rustdoc::clean.

These implementations all emit HTML. As an internal implementation detail, some of them support an alternate format that emits text, but that should not be used external to this module.

Structs

Buffer ğŸ”’

Indent ğŸ”’

WriteCounter ğŸ”’

Enums

Ending ğŸ”’

HrefError ğŸ”’

Traits

Print ğŸ”’

PrintWithSpace ğŸ”’

Functions

anchor ğŸ”’

comma_sep ğŸ”’

display_fn ğŸ”’

fmt_type ğŸ”’

generate_item_def_id_path ğŸ”’

generate_macro_def_id_path ğŸ”’

This function is to get the external macro path because they are not in the cache used in href_with_root_path.

href ğŸ”’

href_relative_parts ğŸ”’

Both paths should only be modules. This is because modules get their own directories; that is, std::vec and std::vec::Vec will both need ../iter/trait.Iterator.html to get at the iterator trait.

href_with_root_path ğŸ”’

join_with_double_colon ğŸ”’

link_tooltip ğŸ”’

make_href ğŸ”’

primitive_link ğŸ”’

primitive_link_fragment ğŸ”’

print_abi_with_space ğŸ”’

print_constness_with_space ğŸ”’

print_default_space ğŸ”’

print_generic_bounds ğŸ”’

print_higher_ranked_params_with_space ğŸ”’

print_where_clause ğŸ”’

The Generics from which to emit a where-clause.The number of spaces to indent each line with.Whether the where-clause needs to add a comma and newline after the last bound.

resolved_path ğŸ”’

Used to render a clean::Path.

to_module_fqp ğŸ”’

tybounds ğŸ”’

url_parts ğŸ”’

visibility_print_with_space ğŸ”’