Module rustdoc::html::format

Unstable (rustdoc)

HTML formatting module

This module contains a large number of fmt::Display implementations for various types in rustdoc::clean. These implementations all currently assume that HTML output is desired, although it may be possible to redesign them in the future to instead emit any format desired.

Structs

AbiSpace 	 [Unstable]

CommaSep 	 [Unstable]
Wrapper struct for emitting a comma-separated list of items

ConstnessSpace 	 [Unstable]
Similarly to VisSpace, this structure is used to render a function constness with a space after it.

HRef 	 [Unstable]

Method 	 [Unstable]
Wrapper struct for properly emitting a method declaration.

MutableSpace 	 [Unstable]
Similar to VisSpace, but used for mutability

RawMutableSpace 	 [Unstable]
Similar to VisSpace, but used for mutability

TyParamBounds 	 [Unstable]
Wrapper struct for emitting type parameter bounds.

UnsafetySpace 	 [Unstable]
Similarly to VisSpace, this structure is used to render a function style with a space after it.

VisSpace 	 [Unstable]
Helper to render an optional visibility with a space after it (if the visibility is preset)

WhereClause 	 [Unstable]
Wrapper struct for emitting a where clause from Generics.

Functions

fmt_impl_for_trait_page 	 [Unstable]

href 	 [Unstable]