Module rustdoc::html::formatUnstable

HTML formatting module

This module contains a large number of fmt::Display implementations for various types in rustdoc::clean. These implementations all currently assume that HTML output is desired, although it may be possible to redesign them in the future to instead emit any format desired.

Structs

CommaSep
Wrapper struct for emitting a comma-separated list of items

ConciseStability
Wrapper struct for emitting the stability level concisely.

Method
Wrapper struct for properly emitting a method declaration.

MutableSpace
Similar to VisSpace, but used for mutability

RawMutableSpace
Similar to VisSpace, but used for mutability

Stability
Wrapper struct for properly emitting the stability level.

TyParamBounds
Wrapper struct for emitting type parameter bounds.

UnsafetySpace
Similarly to VisSpace, this structure is used to render a function style with a space after it.

VisSpace
Helper to render an optional visibility with a space after it (if the visibility is preset)

WhereClause
Wrapper struct for emitting a where clause from Generics.