rustdoc::html::format
Function print_where_clause

pub(crate) fn print_where_clause<'a, 'tcx: 'a>(
    gens: &'a Generics,
    cx: &'a Context<'tcx>,
    indent: usize,
    ending: Ending,
) -> impl Display + 'a + Captures<'tcx>

The Generics from which to emit a where-clause.

The number of spaces to indent each line with.

Whether the where-clause needs to add a comma and newline after the last bound.