Enum rustdoc::html::item_type::ItemType

pub enum ItemType {
    Module,
    ExternCrate,
    Import,
    Struct,
    Enum,
    Function,
    Typedef,
    Static,
    Trait,
    Impl,
    TyMethod,
    Method,
    StructField,
    Variant,
    Macro,
    Primitive,
    AssociatedType,
    Constant,
    AssociatedConst,
}

Unstable (rustdoc)

Item type. Corresponds to clean::ItemEnum variants.

The search index uses item types encoded as smaller numbers which equal to discriminants. JavaScript then is used to decode them into the original value. Consequently, every change to this type should be synchronized to the itemTypes mapping table in static/main.js.

Variants
Module
Unstable (rustdoc)
ExternCrate
Unstable (rustdoc)
Import
Unstable (rustdoc)
Struct
Unstable (rustdoc)
Enum
Unstable (rustdoc)
Function
Unstable (rustdoc)
Typedef
Unstable (rustdoc)
Static
Unstable (rustdoc)
Trait
Unstable (rustdoc)
Impl
Unstable (rustdoc)
TyMethod
Unstable (rustdoc)
Method
Unstable (rustdoc)
StructField
Unstable (rustdoc)
Variant
Unstable (rustdoc)
Macro
Unstable (rustdoc)
Primitive
Unstable (rustdoc)
AssociatedType
Unstable (rustdoc)
Constant
Unstable (rustdoc)
AssociatedConst
Unstable (rustdoc)

Methods

impl ItemType

fn from_item(item: &Item) -> ItemType

fn from_type_kind(kind: TypeKind) -> ItemType

fn to_static_str(&self) -> &'static str