Enum rustdoc::html::item_type::ItemTypeUnstable

pub enum ItemType {
    Module,
    ExternCrate,
    Import,
    Struct,
    Enum,
    Function,
    Typedef,
    Static,
    Trait,
    Impl,
    TyMethod,
    Method,
    StructField,
    Variant,
    Macro,
    Primitive,
    AssociatedType,
    Constant,
}

Item type. Corresponds to clean::ItemEnum variants.

The search index uses item types encoded as smaller numbers which equal to discriminants. JavaScript then is used to decode them into the original value. Consequently, every change to this type should be synchronized to the itemTypes mapping table in static/main.js.

Variants

Module

ExternCrate

Import

Struct

Enum

Function

Typedef

Static

Trait

Impl

TyMethod

Method

StructField

Variant

Macro

Primitive

AssociatedType

Constant

Methods

impl ItemType

fn from_item(item: &Item) -> ItemType

fn from_type_kind(kind: TypeKind) -> ItemType

fn to_static_str(&self) -> &'static str