rustdoc::html
Module markdown

Markdown formatting for rustdoc.

This module implements markdown formatting through the pulldown-cmark library.

#![feature(rustc_private)]

extern crate rustc_span;

use rustc_span::edition::Edition;
use rustdoc::html::markdown::{HeadingOffset, IdMap, Markdown, ErrorCodes};

let s = "My *markdown* _text_";
let mut id_map = IdMap::new();
let md = Markdown {
    content: s,
    links: &[],
    ids: &mut id_map,
    error_codes: ErrorCodes::Yes,
    edition: Edition::Edition2015,
    playground: &None,
    heading_offset: HeadingOffset::H2,
};
let html = md.into_string();
// ... something using html

Modules

footnotes ğŸ”’

Markdown footnote handling.

Structs

CodeBlocks ğŸ”’

Adds syntax highlighting and playground Run buttons to Rust code blocks.

ExtraInfo ğŸ”’

HeadingLinks ğŸ”’

Make headings links with anchor IDs and build up TOC.

IdMap

Indices ğŸ”’

LangString ğŸ”’

LinkReplacer ğŸ”’

Make headings links with anchor IDs and build up TOC.

Markdown

When to_string is called, this struct will emit the HTML corresponding to the rendered version of the contained markdown string.

MarkdownItemInfo ğŸ”’

A tuple struct like Markdown that renders the markdown escaping HTML tags and includes no paragraph tags.

MarkdownLink ğŸ”’

MarkdownSummaryLine ğŸ”’

A tuple struct like Markdown that renders only the first paragraph.

MarkdownWithToc ğŸ”’

A struct like Markdown that renders the markdown with a table of contents.

MdRelLine ğŸ”’

A newtype that represents a relative line number in Markdown.

Playground

RustCodeBlock ğŸ”’

SummaryLine ğŸ”’

Extracts just the first paragraph.

TableWrapper ğŸ”’

Wrap HTML tables into <div> to prevent having the doc blocks width being too big.

TagIterator ğŸ”’

This is the parser for fenced codeblocks attributes. It implements the following eBNF:

Enums

ErrorCodes

HeadingOffset

Ignore ğŸ”’

LangStringToken ğŸ”’

Line ğŸ”’

Controls whether a line will be hidden or shown in HTML output.

MarkdownLinkRange ğŸ”’

Constants

MAX_HEADER_LEVEL ğŸ”’

Functions

check_if_allowed_tag ğŸ”’

find_codes ğŸ”’

find_testable_code ğŸ”’

html_text_from_events ğŸ”’

is_bareword_char ğŸ”’

is_default_id ğŸ”’

is_forbidden_tag ğŸ”’

is_leading_char ğŸ”’

is_separator ğŸ”’

map_line ğŸ”’

markdown_links ğŸ”’

markdown_summary_with_limit ğŸ”’

Renders a subset of Markdown in the first paragraph of the provided Markdown.

plain_text_from_events ğŸ”’

plain_text_summary ğŸ”’

Renders the first paragraph of the provided markdown as plain text. Useful for alt-text.

rust_code_blocks ğŸ”’

Returns a range of bytes for each code block in the markdown that is tagged as rust or untagged (and assumed to be rust).

short_markdown_summary ğŸ”’

Renders a shortened first paragraph of the given Markdown as a subset of Markdown, making it suitable for contexts like the search index.

slugify ğŸ”’

Convert chars from a title for an id.

summary_opts ğŸ”’

Options for rendering Markdown in summaries (e.g., in search results).

main_body_opts ğŸ”’

Options for rendering Markdown in the main body of documentation.

Type Aliases

SpannedEvent ğŸ”’