rustdoc::html::render
Function render_stability_since_raw_with_extra

fn render_stability_since_raw_with_extra(
    w: &mut Buffer,
    stable_version: Option<StableSince>,
    const_stability: Option<ConstStability>,
    extra_class: &str,
) -> bool

Writes a span containing the versions at which an item became stable and/or const-stable. For example, if the item became stable at 1.0.0, and const-stable at 1.45.0, this function would write a span containing “1.0.0 (const: 1.45.0)”.

Returns true if a stability annotation was rendered.

Stability and const-stability are considered separately. If the item is unstable, no version will be written. If the item is const-unstable, “const: unstable” will be appended to the span, with a link to the tracking issue if present. If an item’s stability or const-stability version matches the version of its enclosing item, that version will be omitted.

Note that it is possible for an unstable function to be const-stable. In that case, the span will include the const-stable version, but no stable version will be emitted, as a natural consequence of the above rules.