rustdoc::html::render
Module search_index

Modules

encode ğŸ”’

Structs

SerializedSearchIndex ğŸ”’

The serialized search description sharded version

Enums

SimplifiedParam ğŸ”’

Constants

DESC_INDEX_SHARD_LEN ğŸ”’

Functions

build_index ğŸ”’

Builds the search index from the collected metadata

get_fn_inputs_and_outputs ğŸ”’

Return the full list of types when bounds have been resolved.

get_function_type_for_search ğŸ”’

get_index_type ğŸ”’

get_index_type_id ğŸ”’

simplify_fn_constraint ğŸ”’

simplify_fn_type ğŸ”’

The point of this function is to lower generics and types into the simplified form that the frontend search engine can use.