rustdoc::html::sources
Function print_src

pub(crate) fn print_src(
    writer: impl Write,
    s: &str,
    file_span: Span,
    context: &Context<'_>,
    root_path: &str,
    decoration_info: DecorationInfo,
    source_context: SourceContext<'_>,
)

Wrapper struct to render the source code of a file. This will do things like adding line numbers to the left-hand side.