Struct rustdoc::html::toc::TocBuilder

pub(crate) struct TocBuilder {
    top_level: Toc,
    chain: Vec<TocEntry>,
}

Progressive construction of a table of contents.

Fields
top_level: Tocchain: Vec<TocEntry>

The current hierarchy of parent headings, the levels are strictly increasing (i.e., chain[0].level < chain[1].level < ...) with each entry being the most recent occurrence of a heading with that level (it doesn’t include the most recent occurrences of every level, just, if it is in chain then it is the most recent one).

We also have chain[0].level <= top_level.entries[last].

Implementations

impl TocBuilder

pub(crate) fn new() -> TocBuilder

pub(crate) fn into_toc(self) -> Toc

Converts into a true Toc struct.

fn fold_until(&mut self, level: u32)

Collapse the chain until the first heading more important than level (i.e., lower level)

Example:

## A
# B
# C
## D
## E
### F
#### G
### H

If we are considering H (i.e., level 3), then A and B are in self.top_level, D is in C.children, and C, E, F, G are in self.chain.

When we attempt to push H, we realize that first G is not the parent (level is too high) so it is popped from chain and put into F.children, then F isn’t the parent (level is equal, aka sibling), so it’s also popped and put into E.children.

This leaves us looking at E, which does have a smaller level, and, by construction, it’s the most recent thing with smaller level, i.e., it’s the immediate parent of H.

pub(crate) fn push(&mut self, level: u32, name: String, id: String) -> &str

Push a level level heading into the appropriate place in the hierarchy, returning a string containing the section number in <num>.<num>.<num> format.