rustdoc::html

Module url_parts_builder

Structs

UrlPartsBuilder 🔒

A builder that allows efficiently and easily constructing the part of a URL after the domain: nightly/core/str/struct.Bytes.html.

Constants

AVG_PART_LENGTH 🔒

This is just a guess at the average length of a URL part, used for String::with_capacity calls in the FromIterator and Extend impls.