rustdoc::html::url_parts_builder
Struct UrlPartsBuilder

pub(crate) struct UrlPartsBuilder {
    buf: String,
}

A builder that allows efficiently and easily constructing the part of a URL after the domain: nightly/core/str/struct.Bytes.html.

This type is a wrapper around the final String buffer, but its API is like that of a Vec of URL components.

Fields
buf: String
Implementations

impl UrlPartsBuilder

pub(crate) fn new() -> Self

Create an empty buffer.

fn with_capacity_bytes(count: usize) -> Self

Create an empty buffer with capacity for the specified number of bytes.

pub(crate) fn singleton(part: &str) -> Self

Create a buffer with one URL component.

Examples

Basic usage:

let builder = UrlPartsBuilder::singleton("core");
assert_eq!(builder.finish(), "core");

Adding more components afterward.

let mut builder = UrlPartsBuilder::singleton("core");
builder.push("str");
builder.push_front("nightly");
assert_eq!(builder.finish(), "nightly/core/str");

pub(crate) fn push(&mut self, part: &str)

Push a component onto the buffer.

Examples

Basic usage:

let mut builder = UrlPartsBuilder::new();
builder.push("core");
builder.push("str");
builder.push("struct.Bytes.html");
assert_eq!(builder.finish(), "core/str/struct.Bytes.html");

pub(crate) fn push_fmt(&mut self, args: Arguments<'_>)

Push a component onto the buffer, using format!โs formatting syntax.

Examples

Basic usage (equivalent to the example for UrlPartsBuilder::push):

let mut builder = UrlPartsBuilder::new();
builder.push("core");
builder.push("str");
builder.push_fmt(format_args!("{}.{}.html", "struct", "Bytes"));
assert_eq!(builder.finish(), "core/str/struct.Bytes.html");

pub(crate) fn push_front(&mut self, part: &str)

Push a component onto the front of the buffer.

Examples

Basic usage:

let mut builder = UrlPartsBuilder::new();
builder.push("core");
builder.push("str");
builder.push_front("nightly");
builder.push("struct.Bytes.html");
assert_eq!(builder.finish(), "nightly/core/str/struct.Bytes.html");

pub(crate) fn finish(self) -> String

Get the final String buffer.