Function rustdoc::json::import_finder::get_imports

pub(crate) fn get_imports(krate: Crate) -> (Crate, DefIdSet)

Get the id’s of all items that are pub used in the crate.

We need this to know if a stripped module is pub use mod::*, to decide if it needs to be kept in the index, despite being stripped.

See #100973 and #101103 for times when this information is needed.