rustdoc::passes::collect_intra_doc_links
Function trait_assoc_to_impl_assoc_item

fn trait_assoc_to_impl_assoc_item<'tcx>(
    tcx: TyCtxt<'tcx>,
    impl_id: DefId,
    trait_assoc_id: DefId,
) -> Option<AssocItem>

Find the associated item in the impl impl_id that corresponds to the trait associated item trait_assoc_id.

This function returns None if no associated item was found in the impl. This can occur when the trait associated item has a default value that is not overridden in the impl.

This is just a wrapper around TyCtxt::impl_item_implementor_ids() and TyCtxt::associated_item() (with some helpful logging added).