Struct rustdoc::passes::collect_intra_doc_links::LinkCollector

struct LinkCollector<'a, 'tcx> {
    cx: &'a mut DocContext<'tcx>,
    visited_links: FxHashMap<ResolutionInfo, Option<(Res, Option<UrlFragment>)>>,
}

Fields
cx: &'a mut DocContext<'tcx>visited_links: FxHashMap<ResolutionInfo, Option<(Res, Option<UrlFragment>)>>

Cache the resolved links so we can avoid resolving (and emitting errors for) the same link. The link will be None if it could not be resolved (i.e. the error was cached).

Implementations

impl<'a, 'tcx> LinkCollector<'a, 'tcx>

fn variant_field<'path>( &self, path_str: &'path str, item_id: DefId, module_id: DefId ) -> Result<(Res, DefId), UnresolvedPath<'path>>

Given a full link, parse it as an enum struct variant.

In particular, this will return an error whenever there aren’t three full path segments left in the link.

fn resolve_primitive_associated_item( &self, prim_ty: PrimitiveType, ns: Namespace, item_name: Symbol ) -> Vec<(Res, DefId)>

Given a primitive type, try to resolve an associated item.

fn resolve_self_ty( &self, path_str: &str, ns: Namespace, item_id: DefId ) -> Option<Res>

fn resolve_path( &self, path_str: &str, ns: Namespace, item_id: DefId, module_id: DefId ) -> Option<Res>

Convenience wrapper around doc_link_resolutions.

This also handles resolving true and false as booleans. NOTE: doc_link_resolutions knows only about paths, not about types. Associated items will never be resolved by this function.

fn resolve<'path>( &mut self, path_str: &'path str, ns: Namespace, item_id: DefId, module_id: DefId ) -> Result<Vec<(Res, Option<DefId>)>, UnresolvedPath<'path>>

Resolves a string as a path within a particular namespace. Returns an optional URL fragment in the case of variants and methods.

fn def_id_to_res(&self, ty_id: DefId) -> Option<Res>

Convert a DefId to a Res, where possible.

This is used for resolving type aliases.

fn primitive_type_to_ty(&mut self, prim: PrimitiveType) -> Option<Ty<'tcx>>

Convert a PrimitiveType to a Ty, where possible.

This is used for resolving trait impls for primitives

fn resolve_associated_item( &mut self, root_res: Res, item_name: Symbol, ns: Namespace, module_id: DefId ) -> Vec<(Res, DefId)>

Resolve an associated item, returning its containing page’s Res and the fragment targeting the associated item on its page.

impl LinkCollector<'_, '_>

fn resolve_links(&mut self, item: &Item)

fn resolve_link( &mut self, dox: &String, item: &Item, item_id: DefId, module_id: DefId, PreprocessedMarkdownLink: &PreprocessedMarkdownLink ) -> Option<ItemLink>

This is the entry point for resolving an intra-doc link.

FIXME(jynelson): this is way too many arguments

fn verify_disambiguator( &self, path_str: &str, kind: DefKind, id: DefId, disambiguator: Option<Disambiguator>, item: &Item, diag_info: &DiagnosticInfo<'_> ) -> Option<()>

fn report_disambiguator_mismatch( &self, path_str: &str, specified: Disambiguator, resolved: Res, diag_info: &DiagnosticInfo<'_> )

fn report_rawptr_assoc_feature_gate( &self, dox: &str, ori_link: &MarkdownLinkRange, item: &Item )

fn resolve_with_disambiguator_cached( &mut self, key: ResolutionInfo, diag: DiagnosticInfo<'_>, cache_errors: bool, recoverable: bool ) -> Option<(Res, Option<UrlFragment>)>

fn resolve_with_disambiguator( &mut self, key: &ResolutionInfo, diag: DiagnosticInfo<'_>, recoverable: bool ) -> Vec<(Res, Option<DefId>)>

After parsing the disambiguator, resolve the main part of the link.

fn resolve_display_text( &mut self, explicit_link: &Box<str>, display_res_info: ResolutionInfo, ori_link: &MarkdownLink, diag_info: &DiagnosticInfo<'_> )

Resolve display text if the provided link has separated parts of links.

For example: Inline link [display_text](dest_link) and reference link [display_text][reference_link] has separated parts of links.